Convert Climate Data to 360-Day Years

This function converts climate data to a 360-day calendar, with each month forced to 30 days. Data is either removed or repeated to ensure the adjustment, useful for simplifying climate simulations.

Usage

climate_to_30_day_months(clim_df, start_year = 2014, end_year = 2022)

Arguments

clim_df

A data frame of climate data, which must include columns for date, month, day, anom, temp, and CumulativeRainfall.

start_year

The starting year of the conversion (default is 2014).

end_year

The ending year of the conversion (default is 2022).

Value

A data frame with 360-day years, where each month has exactly 30 days.

Examples

library(lubridate)
# Create example climate data for 1 year
example_dates <- seq(as.Date("2014-01-01"), as.Date("2014-12-31"), by = "day")
clim_df <- data.frame(
  date = example_dates,
  month = month(example_dates),
  day = day(example_dates),
  anom = rnorm(length(example_dates)),
  temp = runif(length(example_dates), min = 20, max = 35),
  CumulativeRainfall = cumsum(runif(length(example_dates), 0, 5))
)
# Run conversion
climate_to_30_day_months(clim_df, start_year = 2014, end_year = 2014)
#>           date         anom     temp    cumrain
#> 1   2014-01-01 -1.400043517 32.55593   2.728617
#> 2   2014-01-02  0.255317055 20.50256   6.397349
#> 3   2014-01-03 -2.437263611 20.64993  11.064141
#> 4   2014-01-04 -0.005571287 31.88154  12.281104
#> 5   2014-01-05  0.621552721 23.04048  13.695899
#> 6   2014-01-06  1.148411606 21.03193  14.779029
#> 7   2014-01-07 -1.821817661 32.96662  15.379236
#> 8   2014-01-08 -0.247325302 21.15635  16.230554
#> 9   2014-01-09 -0.244199607 26.46505  18.838198
#> 10  2014-01-10 -0.282705449 34.60434  23.655669
#> 11  2014-01-11 -0.553699384 28.56436  28.095031
#> 12  2014-01-12  0.628982042 31.31610  31.277264
#> 13  2014-01-13  2.065024895 23.63681  32.924165
#> 14  2014-01-14 -1.630989402 29.40737  35.935278
#> 15  2014-01-15  0.512426950 22.52622  39.796085
#> 16  2014-01-16 -1.863011492 26.67184  40.846762
#> 17  2014-01-17 -0.522012515 22.47013  43.768303
#> 18  2014-01-18 -0.052601910 21.77415  46.478683
#> 19  2014-01-19  0.542996343 25.51197  46.940666
#> 20  2014-01-20 -0.914074827 34.19838  48.787370
#> 21  2014-01-21  0.468154420 33.13116  53.549517
#> 22  2014-01-22  0.362951256 23.90088  58.297963
#> 23  2014-01-23 -1.304543545 29.85948  60.546847
#> 24  2014-01-24  0.737776321 21.82797  61.845939
#> 25  2014-01-25  1.888504929 24.78272  61.940469
#> 26  2014-01-26 -0.097445104 29.34592  65.547495
#> 27  2014-01-27 -0.935847354 26.70496  70.311543
#> 28  2014-01-28 -0.015950311 30.62440  72.512659
#> 29  2014-01-29 -0.826788954 33.35043  75.032721
#> 30  2014-01-30 -1.512399651 21.56460  75.250506
#> 31  2014-02-01  0.176488611 32.62338  84.481826
#> 32  2014-02-02  0.243685465 25.46394  87.510065
#> 33  2014-02-03  1.623548883 29.34384  87.530475
#> 34  2014-02-04  0.112038083 26.91121  90.814757
#> 35  2014-02-05 -0.133997013 22.45118  91.569402
#> 36  2014-02-06 -1.910087468 27.70800  92.565205
#> 37  2014-02-07 -0.279237242 22.14900  93.763397
#> 38  2014-02-08 -0.313445978 29.75448  97.071267
#> 39  2014-02-09  1.067307879 22.32942  98.280211
#> 40  2014-02-10  0.070034850 27.62465  98.637959
#> 41  2014-02-11 -0.639123324 21.28806 102.161319
#> 42  2014-02-12 -0.049964899 27.54540 103.738304
#> 43  2014-02-13 -0.251483443 20.57858 105.617680
#> 44  2014-02-14  0.444797116 32.36049 109.693168
#> 45  2014-02-15  2.755417575 24.28978 110.695968
#> 46  2014-02-16  0.046531380 23.70126 112.205619
#> 47  2014-02-17  0.577709069 25.96990 112.644220
#> 48  2014-02-18  0.118194874 29.48139 113.560821
#> 49  2014-02-19 -1.911720491 34.22868 114.594830
#> 50  2014-02-20  0.862086482 25.10179 118.869628
#> 51  2014-02-21 -0.243236740 29.41049 120.184177
#> 52  2014-02-22 -0.206087195 32.37373 124.910849
#> 53  2014-02-23  0.019177592 22.17673 128.874526
#> 54  2014-02-24  0.029560754 34.50709 133.032435
#> 55  2014-02-25  0.549827542 21.29058 136.127322
#> 56  2014-02-26 -2.274114857 23.60692 138.345990
#> 57  2014-02-27  2.682557184 29.58326 143.169118
#> 58  2014-02-28 -0.361221255 27.55093 143.980111
#> 59  2014-02-29 -0.361221255 27.55093 143.980111
#> 60  2014-02-30 -0.361221255 27.55093 143.980111
#> 61  2014-03-01  0.213355750 23.75785 144.908380
#> 62  2014-03-02  1.074345882 34.68551 147.973887
#> 63  2014-03-03 -0.665088249 29.84869 150.381310
#> 64  2014-03-04  1.113952419 21.34773 152.361999
#> 65  2014-03-05 -0.245896412 29.45814 153.955781
#> 66  2014-03-06 -1.177563309 33.15151 158.147656
#> 67  2014-03-07 -0.975850616 33.95422 159.643674
#> 68  2014-03-08  1.065057320 26.29229 162.723983
#> 69  2014-03-09  0.131670635 27.43340 165.034502
#> 70  2014-03-10  0.488628809 25.29057 169.730532
#> 71  2014-03-11 -1.699450568 23.73102 169.761989
#> 72  2014-03-12 -1.470736306 34.38092 171.509126
#> 73  2014-03-13  0.284150344 25.39940 174.340483
#> 74  2014-03-14  1.337320413 22.98479 176.985269
#> 75  2014-03-15  0.236696283 30.89428 180.170346
#> 76  2014-03-16  1.318293384 22.52199 181.924325
#> 77  2014-03-17  0.523909788 23.39722 186.545181
#> 78  2014-03-18  0.606748047 32.68310 188.915387
#> 79  2014-03-19 -0.109935672 25.18780 192.806058
#> 80  2014-03-20  0.172181715 21.00980 196.799740
#> 81  2014-03-21 -0.090327287 20.40529 201.192240
#> 82  2014-03-22  1.924343341 21.77067 205.741459
#> 83  2014-03-23  1.298392759 20.20543 207.969148
#> 84  2014-03-24  0.748791268 31.03594 209.416945
#> 85  2014-03-25  0.556224329 33.85941 210.933498
#> 86  2014-03-26 -0.548257264 34.73286 215.528620
#> 87  2014-03-27  1.110534893 26.07905 220.207182
#> 88  2014-03-28 -2.612334333 24.04542 223.565424
#> 89  2014-03-29 -0.155693776 27.32433 225.421774
#> 90  2014-03-30  0.433889790 20.76594 228.257431
#> 91  2014-04-01  0.424187575 27.43752 233.489167
#> 92  2014-04-02  1.063101996 23.65687 234.921861
#> 93  2014-04-03  1.048712620 23.83624 238.719857
#> 94  2014-04-04 -0.038102895 28.28478 239.862186
#> 95  2014-04-05  0.486148920 23.77366 241.994713
#> 96  2014-04-06  1.672882611 29.42675 244.277636
#> 97  2014-04-07 -0.354361164 24.74173 247.827170
#> 98  2014-04-08  0.946347886 33.86781 249.264455
#> 99  2014-04-09  1.316826356 33.59738 253.129756
#> 100 2014-04-10 -0.296640025 21.37955 256.002568
#> 101 2014-04-11 -0.387213575 31.06693 256.111546
#> 102 2014-04-12 -0.785432656 28.54611 259.976679
#> 103 2014-04-13 -1.056736867 30.44733 260.272323
#> 104 2014-04-14 -0.795541430 30.51151 260.410565
#> 105 2014-04-15 -1.756275428 20.59911 264.114798
#> 106 2014-04-16 -0.690537897 26.47284 267.081402
#> 107 2014-04-17 -0.558541994 32.21347 270.514786
#> 108 2014-04-18 -0.536663326 28.49879 274.191337
#> 109 2014-04-19  0.227127133 28.82371 276.496863
#> 110 2014-04-20  0.978454920 33.20464 279.806859
#> 111 2014-04-21 -0.208882651 30.78977 279.811649
#> 112 2014-04-22 -1.399410460 34.04643 284.691477
#> 113 2014-04-23  0.258537288 26.44147 285.792294
#> 114 2014-04-24 -0.441799453 24.29281 287.323933
#> 115 2014-04-25  0.568599861 31.55965 290.981766
#> 116 2014-04-26  2.126850459 29.24748 291.649297
#> 117 2014-04-27  0.424858441 22.38912 292.148478
#> 118 2014-04-28 -1.684281532 26.02397 293.005404
#> 119 2014-04-29  0.249401784 20.02470 296.091311
#> 120 2014-04-30  1.072838252 26.26949 299.507628
#> 121 2014-05-01  2.039369263 31.45902 303.460498
#> 122 2014-05-02  0.449453778 31.91717 304.526739
#> 123 2014-05-03  1.391814046 23.63881 308.486793
#> 124 2014-05-04  0.426566547 31.38756 311.449965
#> 125 2014-05-05  0.107583992 20.43656 314.481283
#> 126 2014-05-06  0.022294733 26.56011 315.035126
#> 127 2014-05-07  0.603611011 27.95614 316.013169
#> 128 2014-05-08 -0.262650573 26.44747 316.048523
#> 129 2014-05-09 -0.528264082 32.13945 320.752876
#> 130 2014-05-10  0.192149422 33.83847 321.093923
#> 131 2014-05-11 -1.146199669 30.12535 321.778569
#> 132 2014-05-12  0.846184665 21.40952 323.442654
#> 133 2014-05-13  0.081719629 22.96195 325.891095
#> 134 2014-05-14 -1.305117010 27.13318 326.755868
#> 135 2014-05-15 -0.944912060 30.71557 327.080564
#> 136 2014-05-16  0.454341594 28.07119 329.338899
#> 137 2014-05-17 -0.855202501 33.13075 331.235911
#> 138 2014-05-18 -0.286895219 24.40378 335.782403
#> 139 2014-05-19  0.894961626 23.37148 335.932135
#> 140 2014-05-20  0.067304440 30.36377 336.906117
#> 141 2014-05-21 -0.162676337 24.68494 340.303514
#> 142 2014-05-22 -0.827310169 27.24947 344.040845
#> 143 2014-05-23  1.876505621 21.82695 347.288534
#> 144 2014-05-24  0.766440199 30.35011 347.676437
#> 145 2014-05-25  0.979956696 24.73597 350.674706
#> 146 2014-05-26  1.321780992 32.24282 355.069760
#> 147 2014-05-27 -1.119710829 28.18873 358.729946
#> 148 2014-05-28  0.514599819 34.20976 359.130211
#> 149 2014-05-29 -1.509099836 20.57790 363.293960
#> 150 2014-05-30  1.532741480 29.71466 367.617488
#> 151 2014-06-01  0.122103414 26.49396 374.969349
#> 152 2014-06-02 -1.138012401 31.41340 375.581518
#> 153 2014-06-03 -0.558015129 24.40638 376.550008
#> 154 2014-06-04  1.052538537 21.60998 378.120899
#> 155 2014-06-05  0.677683644 21.09492 378.808564
#> 156 2014-06-06  0.038499547 25.34472 383.047579
#> 157 2014-06-07 -0.356381187 28.57341 385.473689
#> 158 2014-06-08  0.782844102 30.16982 387.824902
#> 159 2014-06-09  0.804411616 30.36079 391.780129
#> 160 2014-06-10 -1.900060823 25.60116 393.345045
#> 161 2014-06-11  0.935784286 21.46967 394.888868
#> 162 2014-06-12 -0.309051503 21.48515 396.614645
#> 163 2014-06-13  0.263066677 31.62202 400.756861
#> 164 2014-06-14 -1.790591856 22.27243 402.126839
#> 165 2014-06-15 -0.788258845 25.91910 404.281019
#> 166 2014-06-16 -1.133021669 34.02498 406.365508
#> 167 2014-06-17  0.363652568 31.09384 407.037069
#> 168 2014-06-18 -0.285887914 29.53502 410.468776
#> 169 2014-06-19  0.517669134 25.07330 411.329167
#> 170 2014-06-20 -0.102908670 34.43624 415.946634
#> 171 2014-06-21 -0.974069593 21.67685 419.015352
#> 172 2014-06-22  1.270672301 29.75728 420.593499
#> 173 2014-06-23  0.960864787 34.75267 424.642165
#> 174 2014-06-24  0.768721370 22.68691 429.154035
#> 175 2014-06-25  1.035930771 22.06956 429.780524
#> 176 2014-06-26 -0.473887074 20.84838 432.608310
#> 177 2014-06-27 -1.275334875 28.19551 433.246516
#> 178 2014-06-28 -0.305620674 26.99757 436.384504
#> 179 2014-06-29  2.211769487 33.44568 436.761124
#> 180 2014-06-30 -1.041668381 20.27797 438.188380
#> 181 2014-07-01 -1.146523850 32.76461 439.741392
#> 182 2014-07-02 -1.675327303 27.52135 440.475030
#> 183 2014-07-03  1.525938655 22.49243 444.473883
#> 184 2014-07-04  0.554185515 25.41509 445.333863
#> 185 2014-07-05  1.993110265 29.78000 449.262282
#> 186 2014-07-06 -0.154120740 33.11265 451.612461
#> 187 2014-07-07  2.564408338 30.92458 453.777250
#> 188 2014-07-08  1.061999145 26.18709 456.649914
#> 189 2014-07-09  1.142694878 22.38844 460.299764
#> 190 2014-07-10  1.123838843 32.68555 464.770700
#> 191 2014-07-11 -0.397001493 33.26554 465.872231
#> 192 2014-07-12 -0.823261151 25.93997 467.330214
#> 193 2014-07-13 -0.578884625 20.57211 471.394707
#> 194 2014-07-14  1.763789378 34.27094 471.612719
#> 195 2014-07-15  0.132992146 31.75591 474.139077
#> 196 2014-07-16  0.376499328 23.29160 475.347357
#> 197 2014-07-17  1.138707653 24.96398 478.293069
#> 198 2014-07-18  1.241263075 22.66639 483.148772
#> 199 2014-07-19  0.612090945 25.43774 484.312377
#> 200 2014-07-20 -0.429380087 25.87180 488.649467
#> 201 2014-07-21  1.360461327 33.55816 489.203275
#> 202 2014-07-22 -0.070857431 24.14685 493.210734
#> 203 2014-07-23 -0.272153684 22.09039 496.500357
#> 204 2014-07-24 -2.446680029 26.77004 499.760628
#> 205 2014-07-25  0.065486641 21.65495 500.892241
#> 206 2014-07-26 -1.098508902 33.21435 505.162139
#> 207 2014-07-27 -0.633178176 28.00764 507.340175
#> 208 2014-07-28 -2.063654451 31.08876 508.786583
#> 209 2014-07-29  2.648932029 30.79284 511.974507
#> 210 2014-07-30 -1.153398386 23.96723 513.297204
#> 211 2014-08-01  0.786362576 23.14432 518.324516
#> 212 2014-08-02 -1.270513110 30.35608 521.032906
#> 213 2014-08-03  0.542141549 22.92139 523.457061
#> 214 2014-08-04  0.075105900 31.73231 527.278536
#> 215 2014-08-05  0.558514422 20.22548 532.083670
#> 216 2014-08-06  0.415406399 34.01259 533.436593
#> 217 2014-08-07 -1.452299769 21.26654 537.089594
#> 218 2014-08-08  0.941206122 20.51078 537.760422
#> 219 2014-08-09 -0.338935872 24.66439 538.002413
#> 220 2014-08-10 -0.075574247 31.77065 541.520448
#> 221 2014-08-11  0.040204392 24.20601 542.256517
#> 222 2014-08-12  0.124301066 22.47926 546.102294
#> 223 2014-08-13 -0.998432551 26.50392 546.984045
#> 224 2014-08-14  1.233390065 32.21068 547.537840
#> 225 2014-08-15  0.340424488 30.29771 552.229552
#> 226 2014-08-16 -0.472702482 31.65959 556.461907
#> 227 2014-08-17  0.708753061 21.36801 559.317625
#> 228 2014-08-18 -1.528958715 31.13936 562.713092
#> 229 2014-08-19  0.237425345 25.95011 563.159704
#> 230 2014-08-20 -1.312814246 24.44042 567.086574
#> 231 2014-08-21  0.747028587 31.14386 568.221361
#> 232 2014-08-22 -1.562518435 26.99764 570.462254
#> 233 2014-08-23  0.071053360 31.59542 571.268392
#> 234 2014-08-24 -0.639534770 20.33176 572.148950
#> 235 2014-08-25 -0.845195739 32.22535 573.140091
#> 236 2014-08-26  0.675244698 24.55750 574.928148
#> 237 2014-08-27  1.153375794 33.97694 575.834814
#> 238 2014-08-28 -1.686504742 28.10647 578.640538
#> 239 2014-08-29 -0.902814949 23.82228 581.935953
#> 240 2014-08-30  1.317633698 27.16381 585.237220
#> 241 2014-09-01  1.203767839 20.11258 590.216454
#> 242 2014-09-02 -1.431270777 31.58613 593.353900
#> 243 2014-09-03  1.382910861 23.98690 593.427052
#> 244 2014-09-04  0.003125940 34.56561 594.452943
#> 245 2014-09-05 -0.077886824 22.85788 597.768326
#> 246 2014-09-06  0.441428226 20.29544 600.087046
#> 247 2014-09-07  0.128922896 29.18565 601.888787
#> 248 2014-09-08 -0.830214260 27.88597 605.335619
#> 249 2014-09-09 -0.503592910 29.51780 606.629923
#> 250 2014-09-10 -1.193641182 25.93172 610.885948
#> 251 2014-09-11 -0.751723323 22.62302 613.285972
#> 252 2014-09-12  1.455841403 24.90970 615.972304
#> 253 2014-09-13 -0.828603533 29.22130 619.280503
#> 254 2014-09-14  0.289774460 21.18750 620.784776
#> 255 2014-09-15 -0.480053484 27.78591 622.102283
#> 256 2014-09-16 -0.604829354 24.59422 623.627514
#> 257 2014-09-17  1.460110180 30.07093 628.002305
#> 258 2014-09-18  0.149679354 25.90508 631.594491
#> 259 2014-09-19 -1.433321100 31.21772 633.629090
#> 260 2014-09-20 -0.010303319 22.08390 633.725887
#> 261 2014-09-21 -0.212236035 25.66015 633.976789
#> 262 2014-09-22 -0.906340179 29.62075 637.488044
#> 263 2014-09-23 -2.102152479 34.91292 637.819896
#> 264 2014-09-24  1.893360464 25.52839 638.000557
#> 265 2014-09-25 -0.968125837 25.06023 638.101326
#> 266 2014-09-26 -0.102603036 33.70528 639.965231
#> 267 2014-09-27  0.239959572 33.64131 641.093558
#> 268 2014-09-28  0.060898893 23.57570 643.553149
#> 269 2014-09-29 -2.177576028 26.17520 645.786375
#> 270 2014-09-30 -0.117860143 31.19108 649.237399
#> 271 2014-10-01  0.112294787 30.85176 653.451099
#> 272 2014-10-02  0.007886198 29.21786 655.330267
#> 273 2014-10-03  1.877743872 27.10971 660.254716
#> 274 2014-10-04  2.158756554 28.67692 665.016847
#> 275 2014-10-05  0.709714522 23.62391 669.966340
#> 276 2014-10-06  0.766983379 21.71568 672.181982
#> 277 2014-10-07 -0.308211421 24.47517 672.950801
#> 278 2014-10-08  1.012001849 27.26179 677.674438
#> 279 2014-10-09 -0.919051597 28.51109 680.271770
#> 280 2014-10-10  0.563380077 31.30930 682.549943
#> 281 2014-10-11  0.322482749 28.18138 683.650394
#> 282 2014-10-12  0.366674363 31.39517 684.346306
#> 283 2014-10-13  1.129835153 23.19451 685.427095
#> 284 2014-10-14 -0.941498076 32.58898 687.455488
#> 285 2014-10-15  0.217837643 26.10046 688.624161
#> 286 2014-10-16  1.415412293 33.10610 688.749493
#> 287 2014-10-17 -0.383733048 33.23235 692.511203
#> 288 2014-10-18 -0.174086374 21.61534 696.171566
#> 289 2014-10-19 -0.221744517 28.61817 698.217100
#> 290 2014-10-20 -1.009528722 34.94050 701.371389
#> 291 2014-10-21  0.480725266 32.75830 701.698613
#> 292 2014-10-22  1.604407328 26.56483 702.442074
#> 293 2014-10-23 -1.515024529 24.89645 703.840583
#> 294 2014-10-24 -1.416023914 25.03953 708.093867
#> 295 2014-10-25  0.876777327 23.58911 708.201571
#> 296 2014-10-26  0.624132413 26.32057 709.465720
#> 297 2014-10-27  2.112277288 27.68205 712.749278
#> 298 2014-10-28 -0.356124416 22.78170 716.695025
#> 299 2014-10-29 -1.064464209 24.89601 720.239253
#> 300 2014-10-30  1.077116538 32.25610 720.562218
#> 301 2014-11-01  0.198392095 27.97613 724.123417
#> 302 2014-11-02 -0.400405249 34.42357 728.193995
#> 303 2014-11-03  0.616154281 21.56070 731.462895
#> 304 2014-11-04  1.974156748 31.05685 733.229155
#> 305 2014-11-05  1.884662324 32.71220 736.212129
#> 306 2014-11-06 -1.588620547 29.65608 739.850651
#> 307 2014-11-07 -0.539923164 23.48686 744.828494
#> 308 2014-11-08 -1.169461464 33.74344 747.318770
#> 309 2014-11-09  0.559105989 26.36896 748.470244
#> 310 2014-11-10 -1.819347247 22.89415 749.976516
#> 311 2014-11-11  0.393343972 30.52326 753.578849
#> 312 2014-11-12  0.042134106 34.88871 754.885876
#> 313 2014-11-13  1.179664177 30.63369 755.424358
#> 314 2014-11-14 -0.256921176 27.76228 756.736471
#> 315 2014-11-15 -1.056336098 30.63098 757.670002
#> 316 2014-11-16  0.198777205 25.62711 760.382211
#> 317 2014-11-17  0.650533552 23.82161 760.709803
#> 318 2014-11-18  0.343913337 34.94619 765.111729
#> 319 2014-11-19  1.477532312 27.84343 766.507792
#> 320 2014-11-20  0.072025698 30.73714 769.871310
#> 321 2014-11-21  2.126444534 20.03739 770.171964
#> 322 2014-11-22 -1.476196906 24.39453 775.114392
#> 323 2014-11-23  0.407888500 22.71224 776.525769
#> 324 2014-11-24  1.393977798 26.11391 779.296331
#> 325 2014-11-25  0.360278296 23.25494 783.746751
#> 326 2014-11-26  0.654550251 21.07014 783.860761
#> 327 2014-11-27  1.052155422 23.79967 787.249204
#> 328 2014-11-28 -1.979555125 29.78719 788.934503
#> 329 2014-11-29  1.208385605 31.01666 789.640483
#> 330 2014-11-30 -0.169280084 32.26270 791.403303
#> 331 2014-12-01  0.295029753 24.58986 793.410527
#> 332 2014-12-02  1.266340587 33.27871 797.855099
#> 333 2014-12-03 -1.135343257 29.09838 799.418170
#> 334 2014-12-04 -1.131053798 33.21529 799.726443
#> 335 2014-12-05  0.109993384 30.19881 803.123875
#> 336 2014-12-06  0.852905410 22.14415 805.014688
#> 337 2014-12-07 -0.234337862 31.48198 807.956539
#> 338 2014-12-08  2.086688567 29.40450 810.416432
#> 339 2014-12-09 -0.110919371 30.95508 811.845317
#> 340 2014-12-10 -1.392847056 26.35557 811.887415
#> 341 2014-12-11 -1.142290768 24.02496 814.641286
#> 342 2014-12-12  1.704608737 31.53572 818.707441
#> 343 2014-12-13 -0.080073634 28.81020 821.925314
#> 344 2014-12-14 -0.437281240 33.47852 822.270624
#> 345 2014-12-15 -0.119215094 33.05487 825.914205
#> 346 2014-12-16  0.786462865 25.98959 829.274761
#> 347 2014-12-17 -0.578945246 34.47679 833.373766
#> 348 2014-12-18 -0.145426885 27.43987 837.946874
#> 349 2014-12-19  0.526457991 26.99896 841.396727
#> 350 2014-12-20  1.733578110 21.61617 843.856616
#> 351 2014-12-21  1.448657220 33.71447 847.091259
#> 352 2014-12-22  1.518193149 22.15705 849.373834
#> 353 2014-12-23 -0.384007254 23.97781 850.001067
#> 354 2014-12-24  1.827125177 22.22695 853.022892
#> 355 2014-12-25 -0.551491750 26.21373 857.157026
#> 356 2014-12-26 -0.865753541 21.63425 860.253917
#> 357 2014-12-27 -0.343831481 23.70617 864.459186
#> 358 2014-12-28  1.062876458 28.85965 866.595419
#> 359 2014-12-29  0.813058204 24.92401 870.441852
#> 360 2014-12-30  1.803483361 30.99456 874.002556