Compare Mean Bias for 2019 vs Other Years

This function compares the bias of simulated data against observed data for the year 2019 (SMC) versus other years (non-SMC).

Usage

compare_bias_2019_vs_other_years(simulated_df, obs_cases)

Arguments

simulated_df

A data frame containing the simulated data with a date_ymd column and metrics such as inc_A, inc_C, and inc.

obs_cases

A data frame containing the observed data with a date_ymd column and metrics such as inc_A, inc_C, and inc.

Value

A list containing:

smc_bias: A data frame with total bias for 2019 (SMC).

mean_non_smc_bias: A data frame summarizing the mean bias for non-SMC years across all metrics.

Examples

simulated_df <- data.frame(
  date_ymd = as.Date(c("2018-01-01", "2019-01-01", "2019-02-01")),
  inc_A = c(5, 4, 6),
  inc_C = c(2, 3, 5),
  inc = c(7, 6, 11)
)
obs_cases <- data.frame(
  date_ymd = as.Date(c("2018-01-01", "2019-01-01", "2019-02-01")),
  inc_A = c(6, 5, 7),
  inc_C = c(3, 4, 6),
  inc = c(9, 9, 13)
)
compare_bias_2019_vs_other_years(simulated_df, obs_cases)
#> $smc_bias
#> # A tibble: 1 × 4
#>    year total_bias_inc_A total_bias_inc_C total_bias_inc
#>   <dbl>            <dbl>            <dbl>          <dbl>
#> 1  2019                2                2              5
#> 
#> $mean_non_smc_bias
#> # A tibble: 1 × 3
#>   mean_bias_inc_A mean_bias_inc_C mean_bias_inc
#>             <dbl>           <dbl>         <dbl>
#> 1               1               1             2
#>