Generate Dates for a 360-Day Year Calendar

This function generates a sequence of dates assuming a 360-day year, with each month consisting of 30 days. Useful for simplified models where leap years and 31-day months are ignored.

Usage

generate_360_day_dates(start_year, end_year)

Arguments

start_year

The starting year of the date sequence.

end_year

The ending year of the date sequence.

Value

A vector of dates formatted as "YYYY-MM-DD" for each day within the given range, assuming a 360-day year.

Examples

generate_360_day_dates(2014, 2022)
#>    [1] "2014-01-01" "2014-01-02" "2014-01-03" "2014-01-04" "2014-01-05"
#>    [6] "2014-01-06" "2014-01-07" "2014-01-08" "2014-01-09" "2014-01-10"
#>   [11] "2014-01-11" "2014-01-12" "2014-01-13" "2014-01-14" "2014-01-15"
#>   [16] "2014-01-16" "2014-01-17" "2014-01-18" "2014-01-19" "2014-01-20"
#>   [21] "2014-01-21" "2014-01-22" "2014-01-23" "2014-01-24" "2014-01-25"
#>   [26] "2014-01-26" "2014-01-27" "2014-01-28" "2014-01-29" "2014-01-30"
#>   [31] "2014-02-01" "2014-02-02" "2014-02-03" "2014-02-04" "2014-02-05"
#>   [36] "2014-02-06" "2014-02-07" "2014-02-08" "2014-02-09" "2014-02-10"
#>   [41] "2014-02-11" "2014-02-12" "2014-02-13" "2014-02-14" "2014-02-15"
#>   [46] "2014-02-16" "2014-02-17" "2014-02-18" "2014-02-19" "2014-02-20"
#>   [51] "2014-02-21" "2014-02-22" "2014-02-23" "2014-02-24" "2014-02-25"
#>   [56] "2014-02-26" "2014-02-27" "2014-02-28" "2014-02-29" "2014-02-30"
#>   [61] "2014-03-01" "2014-03-02" "2014-03-03" "2014-03-04" "2014-03-05"
#>   [66] "2014-03-06" "2014-03-07" "2014-03-08" "2014-03-09" "2014-03-10"
#>   [71] "2014-03-11" "2014-03-12" "2014-03-13" "2014-03-14" "2014-03-15"
#>   [76] "2014-03-16" "2014-03-17" "2014-03-18" "2014-03-19" "2014-03-20"
#>   [81] "2014-03-21" "2014-03-22" "2014-03-23" "2014-03-24" "2014-03-25"
#>   [86] "2014-03-26" "2014-03-27" "2014-03-28" "2014-03-29" "2014-03-30"
#>   [91] "2014-04-01" "2014-04-02" "2014-04-03" "2014-04-04" "2014-04-05"
#>   [96] "2014-04-06" "2014-04-07" "2014-04-08" "2014-04-09" "2014-04-10"
#>  [101] "2014-04-11" "2014-04-12" "2014-04-13" "2014-04-14" "2014-04-15"
#>  [106] "2014-04-16" "2014-04-17" "2014-04-18" "2014-04-19" "2014-04-20"
#>  [111] "2014-04-21" "2014-04-22" "2014-04-23" "2014-04-24" "2014-04-25"
#>  [116] "2014-04-26" "2014-04-27" "2014-04-28" "2014-04-29" "2014-04-30"
#>  [121] "2014-05-01" "2014-05-02" "2014-05-03" "2014-05-04" "2014-05-05"
#>  [126] "2014-05-06" "2014-05-07" "2014-05-08" "2014-05-09" "2014-05-10"
#>  [131] "2014-05-11" "2014-05-12" "2014-05-13" "2014-05-14" "2014-05-15"
#>  [136] "2014-05-16" "2014-05-17" "2014-05-18" "2014-05-19" "2014-05-20"
#>  [141] "2014-05-21" "2014-05-22" "2014-05-23" "2014-05-24" "2014-05-25"
#>  [146] "2014-05-26" "2014-05-27" "2014-05-28" "2014-05-29" "2014-05-30"
#>  [151] "2014-06-01" "2014-06-02" "2014-06-03" "2014-06-04" "2014-06-05"
#>  [156] "2014-06-06" "2014-06-07" "2014-06-08" "2014-06-09" "2014-06-10"
#>  [161] "2014-06-11" "2014-06-12" "2014-06-13" "2014-06-14" "2014-06-15"
#>  [166] "2014-06-16" "2014-06-17" "2014-06-18" "2014-06-19" "2014-06-20"
#>  [171] "2014-06-21" "2014-06-22" "2014-06-23" "2014-06-24" "2014-06-25"
#>  [176] "2014-06-26" "2014-06-27" "2014-06-28" "2014-06-29" "2014-06-30"
#>  [181] "2014-07-01" "2014-07-02" "2014-07-03" "2014-07-04" "2014-07-05"
#>  [186] "2014-07-06" "2014-07-07" "2014-07-08" "2014-07-09" "2014-07-10"
#>  [191] "2014-07-11" "2014-07-12" "2014-07-13" "2014-07-14" "2014-07-15"
#>  [196] "2014-07-16" "2014-07-17" "2014-07-18" "2014-07-19" "2014-07-20"
#>  [201] "2014-07-21" "2014-07-22" "2014-07-23" "2014-07-24" "2014-07-25"
#>  [206] "2014-07-26" "2014-07-27" "2014-07-28" "2014-07-29" "2014-07-30"
#>  [211] "2014-08-01" "2014-08-02" "2014-08-03" "2014-08-04" "2014-08-05"
#>  [216] "2014-08-06" "2014-08-07" "2014-08-08" "2014-08-09" "2014-08-10"
#>  [221] "2014-08-11" "2014-08-12" "2014-08-13" "2014-08-14" "2014-08-15"
#>  [226] "2014-08-16" "2014-08-17" "2014-08-18" "2014-08-19" "2014-08-20"
#>  [231] "2014-08-21" "2014-08-22" "2014-08-23" "2014-08-24" "2014-08-25"
#>  [236] "2014-08-26" "2014-08-27" "2014-08-28" "2014-08-29" "2014-08-30"
#>  [241] "2014-09-01" "2014-09-02" "2014-09-03" "2014-09-04" "2014-09-05"
#>  [246] "2014-09-06" "2014-09-07" "2014-09-08" "2014-09-09" "2014-09-10"
#>  [251] "2014-09-11" "2014-09-12" "2014-09-13" "2014-09-14" "2014-09-15"
#>  [256] "2014-09-16" "2014-09-17" "2014-09-18" "2014-09-19" "2014-09-20"
#>  [261] "2014-09-21" "2014-09-22" "2014-09-23" "2014-09-24" "2014-09-25"
#>  [266] "2014-09-26" "2014-09-27" "2014-09-28" "2014-09-29" "2014-09-30"
#>  [271] "2014-10-01" "2014-10-02" "2014-10-03" "2014-10-04" "2014-10-05"
#>  [276] "2014-10-06" "2014-10-07" "2014-10-08" "2014-10-09" "2014-10-10"
#>  [281] "2014-10-11" "2014-10-12" "2014-10-13" "2014-10-14" "2014-10-15"
#>  [286] "2014-10-16" "2014-10-17" "2014-10-18" "2014-10-19" "2014-10-20"
#>  [291] "2014-10-21" "2014-10-22" "2014-10-23" "2014-10-24" "2014-10-25"
#>  [296] "2014-10-26" "2014-10-27" "2014-10-28" "2014-10-29" "2014-10-30"
#>  [301] "2014-11-01" "2014-11-02" "2014-11-03" "2014-11-04" "2014-11-05"
#>  [306] "2014-11-06" "2014-11-07" "2014-11-08" "2014-11-09" "2014-11-10"
#>  [311] "2014-11-11" "2014-11-12" "2014-11-13" "2014-11-14" "2014-11-15"
#>  [316] "2014-11-16" "2014-11-17" "2014-11-18" "2014-11-19" "2014-11-20"
#>  [321] "2014-11-21" "2014-11-22" "2014-11-23" "2014-11-24" "2014-11-25"
#>  [326] "2014-11-26" "2014-11-27" "2014-11-28" "2014-11-29" "2014-11-30"
#>  [331] "2014-12-01" "2014-12-02" "2014-12-03" "2014-12-04" "2014-12-05"
#>  [336] "2014-12-06" "2014-12-07" "2014-12-08" "2014-12-09" "2014-12-10"
#>  [341] "2014-12-11" "2014-12-12" "2014-12-13" "2014-12-14" "2014-12-15"
#>  [346] "2014-12-16" "2014-12-17" "2014-12-18" "2014-12-19" "2014-12-20"
#>  [351] "2014-12-21" "2014-12-22" "2014-12-23" "2014-12-24" "2014-12-25"
#>  [356] "2014-12-26" "2014-12-27" "2014-12-28" "2014-12-29" "2014-12-30"
#>  [361] "2015-01-01" "2015-01-02" "2015-01-03" "2015-01-04" "2015-01-05"
#>  [366] "2015-01-06" "2015-01-07" "2015-01-08" "2015-01-09" "2015-01-10"
#>  [371] "2015-01-11" "2015-01-12" "2015-01-13" "2015-01-14" "2015-01-15"
#>  [376] "2015-01-16" "2015-01-17" "2015-01-18" "2015-01-19" "2015-01-20"
#>  [381] "2015-01-21" "2015-01-22" "2015-01-23" "2015-01-24" "2015-01-25"
#>  [386] "2015-01-26" "2015-01-27" "2015-01-28" "2015-01-29" "2015-01-30"
#>  [391] "2015-02-01" "2015-02-02" "2015-02-03" "2015-02-04" "2015-02-05"
#>  [396] "2015-02-06" "2015-02-07" "2015-02-08" "2015-02-09" "2015-02-10"
#>  [401] "2015-02-11" "2015-02-12" "2015-02-13" "2015-02-14" "2015-02-15"
#>  [406] "2015-02-16" "2015-02-17" "2015-02-18" "2015-02-19" "2015-02-20"
#>  [411] "2015-02-21" "2015-02-22" "2015-02-23" "2015-02-24" "2015-02-25"
#>  [416] "2015-02-26" "2015-02-27" "2015-02-28" "2015-02-29" "2015-02-30"
#>  [421] "2015-03-01" "2015-03-02" "2015-03-03" "2015-03-04" "2015-03-05"
#>  [426] "2015-03-06" "2015-03-07" "2015-03-08" "2015-03-09" "2015-03-10"
#>  [431] "2015-03-11" "2015-03-12" "2015-03-13" "2015-03-14" "2015-03-15"
#>  [436] "2015-03-16" "2015-03-17" "2015-03-18" "2015-03-19" "2015-03-20"
#>  [441] "2015-03-21" "2015-03-22" "2015-03-23" "2015-03-24" "2015-03-25"
#>  [446] "2015-03-26" "2015-03-27" "2015-03-28" "2015-03-29" "2015-03-30"
#>  [451] "2015-04-01" "2015-04-02" "2015-04-03" "2015-04-04" "2015-04-05"
#>  [456] "2015-04-06" "2015-04-07" "2015-04-08" "2015-04-09" "2015-04-10"
#>  [461] "2015-04-11" "2015-04-12" "2015-04-13" "2015-04-14" "2015-04-15"
#>  [466] "2015-04-16" "2015-04-17" "2015-04-18" "2015-04-19" "2015-04-20"
#>  [471] "2015-04-21" "2015-04-22" "2015-04-23" "2015-04-24" "2015-04-25"
#>  [476] "2015-04-26" "2015-04-27" "2015-04-28" "2015-04-29" "2015-04-30"
#>  [481] "2015-05-01" "2015-05-02" "2015-05-03" "2015-05-04" "2015-05-05"
#>  [486] "2015-05-06" "2015-05-07" "2015-05-08" "2015-05-09" "2015-05-10"
#>  [491] "2015-05-11" "2015-05-12" "2015-05-13" "2015-05-14" "2015-05-15"
#>  [496] "2015-05-16" "2015-05-17" "2015-05-18" "2015-05-19" "2015-05-20"
#>  [501] "2015-05-21" "2015-05-22" "2015-05-23" "2015-05-24" "2015-05-25"
#>  [506] "2015-05-26" "2015-05-27" "2015-05-28" "2015-05-29" "2015-05-30"
#>  [511] "2015-06-01" "2015-06-02" "2015-06-03" "2015-06-04" "2015-06-05"
#>  [516] "2015-06-06" "2015-06-07" "2015-06-08" "2015-06-09" "2015-06-10"
#>  [521] "2015-06-11" "2015-06-12" "2015-06-13" "2015-06-14" "2015-06-15"
#>  [526] "2015-06-16" "2015-06-17" "2015-06-18" "2015-06-19" "2015-06-20"
#>  [531] "2015-06-21" "2015-06-22" "2015-06-23" "2015-06-24" "2015-06-25"
#>  [536] "2015-06-26" "2015-06-27" "2015-06-28" "2015-06-29" "2015-06-30"
#>  [541] "2015-07-01" "2015-07-02" "2015-07-03" "2015-07-04" "2015-07-05"
#>  [546] "2015-07-06" "2015-07-07" "2015-07-08" "2015-07-09" "2015-07-10"
#>  [551] "2015-07-11" "2015-07-12" "2015-07-13" "2015-07-14" "2015-07-15"
#>  [556] "2015-07-16" "2015-07-17" "2015-07-18" "2015-07-19" "2015-07-20"
#>  [561] "2015-07-21" "2015-07-22" "2015-07-23" "2015-07-24" "2015-07-25"
#>  [566] "2015-07-26" "2015-07-27" "2015-07-28" "2015-07-29" "2015-07-30"
#>  [571] "2015-08-01" "2015-08-02" "2015-08-03" "2015-08-04" "2015-08-05"
#>  [576] "2015-08-06" "2015-08-07" "2015-08-08" "2015-08-09" "2015-08-10"
#>  [581] "2015-08-11" "2015-08-12" "2015-08-13" "2015-08-14" "2015-08-15"
#>  [586] "2015-08-16" "2015-08-17" "2015-08-18" "2015-08-19" "2015-08-20"
#>  [591] "2015-08-21" "2015-08-22" "2015-08-23" "2015-08-24" "2015-08-25"
#>  [596] "2015-08-26" "2015-08-27" "2015-08-28" "2015-08-29" "2015-08-30"
#>  [601] "2015-09-01" "2015-09-02" "2015-09-03" "2015-09-04" "2015-09-05"
#>  [606] "2015-09-06" "2015-09-07" "2015-09-08" "2015-09-09" "2015-09-10"
#>  [611] "2015-09-11" "2015-09-12" "2015-09-13" "2015-09-14" "2015-09-15"
#>  [616] "2015-09-16" "2015-09-17" "2015-09-18" "2015-09-19" "2015-09-20"
#>  [621] "2015-09-21" "2015-09-22" "2015-09-23" "2015-09-24" "2015-09-25"
#>  [626] "2015-09-26" "2015-09-27" "2015-09-28" "2015-09-29" "2015-09-30"
#>  [631] "2015-10-01" "2015-10-02" "2015-10-03" "2015-10-04" "2015-10-05"
#>  [636] "2015-10-06" "2015-10-07" "2015-10-08" "2015-10-09" "2015-10-10"
#>  [641] "2015-10-11" "2015-10-12" "2015-10-13" "2015-10-14" "2015-10-15"
#>  [646] "2015-10-16" "2015-10-17" "2015-10-18" "2015-10-19" "2015-10-20"
#>  [651] "2015-10-21" "2015-10-22" "2015-10-23" "2015-10-24" "2015-10-25"
#>  [656] "2015-10-26" "2015-10-27" "2015-10-28" "2015-10-29" "2015-10-30"
#>  [661] "2015-11-01" "2015-11-02" "2015-11-03" "2015-11-04" "2015-11-05"
#>  [666] "2015-11-06" "2015-11-07" "2015-11-08" "2015-11-09" "2015-11-10"
#>  [671] "2015-11-11" "2015-11-12" "2015-11-13" "2015-11-14" "2015-11-15"
#>  [676] "2015-11-16" "2015-11-17" "2015-11-18" "2015-11-19" "2015-11-20"
#>  [681] "2015-11-21" "2015-11-22" "2015-11-23" "2015-11-24" "2015-11-25"
#>  [686] "2015-11-26" "2015-11-27" "2015-11-28" "2015-11-29" "2015-11-30"
#>  [691] "2015-12-01" "2015-12-02" "2015-12-03" "2015-12-04" "2015-12-05"
#>  [696] "2015-12-06" "2015-12-07" "2015-12-08" "2015-12-09" "2015-12-10"
#>  [701] "2015-12-11" "2015-12-12" "2015-12-13" "2015-12-14" "2015-12-15"
#>  [706] "2015-12-16" "2015-12-17" "2015-12-18" "2015-12-19" "2015-12-20"
#>  [711] "2015-12-21" "2015-12-22" "2015-12-23" "2015-12-24" "2015-12-25"
#>  [716] "2015-12-26" "2015-12-27" "2015-12-28" "2015-12-29" "2015-12-30"
#>  [721] "2016-01-01" "2016-01-02" "2016-01-03" "2016-01-04" "2016-01-05"
#>  [726] "2016-01-06" "2016-01-07" "2016-01-08" "2016-01-09" "2016-01-10"
#>  [731] "2016-01-11" "2016-01-12" "2016-01-13" "2016-01-14" "2016-01-15"
#>  [736] "2016-01-16" "2016-01-17" "2016-01-18" "2016-01-19" "2016-01-20"
#>  [741] "2016-01-21" "2016-01-22" "2016-01-23" "2016-01-24" "2016-01-25"
#>  [746] "2016-01-26" "2016-01-27" "2016-01-28" "2016-01-29" "2016-01-30"
#>  [751] "2016-02-01" "2016-02-02" "2016-02-03" "2016-02-04" "2016-02-05"
#>  [756] "2016-02-06" "2016-02-07" "2016-02-08" "2016-02-09" "2016-02-10"
#>  [761] "2016-02-11" "2016-02-12" "2016-02-13" "2016-02-14" "2016-02-15"
#>  [766] "2016-02-16" "2016-02-17" "2016-02-18" "2016-02-19" "2016-02-20"
#>  [771] "2016-02-21" "2016-02-22" "2016-02-23" "2016-02-24" "2016-02-25"
#>  [776] "2016-02-26" "2016-02-27" "2016-02-28" "2016-02-29" "2016-02-30"
#>  [781] "2016-03-01" "2016-03-02" "2016-03-03" "2016-03-04" "2016-03-05"
#>  [786] "2016-03-06" "2016-03-07" "2016-03-08" "2016-03-09" "2016-03-10"
#>  [791] "2016-03-11" "2016-03-12" "2016-03-13" "2016-03-14" "2016-03-15"
#>  [796] "2016-03-16" "2016-03-17" "2016-03-18" "2016-03-19" "2016-03-20"
#>  [801] "2016-03-21" "2016-03-22" "2016-03-23" "2016-03-24" "2016-03-25"
#>  [806] "2016-03-26" "2016-03-27" "2016-03-28" "2016-03-29" "2016-03-30"
#>  [811] "2016-04-01" "2016-04-02" "2016-04-03" "2016-04-04" "2016-04-05"
#>  [816] "2016-04-06" "2016-04-07" "2016-04-08" "2016-04-09" "2016-04-10"
#>  [821] "2016-04-11" "2016-04-12" "2016-04-13" "2016-04-14" "2016-04-15"
#>  [826] "2016-04-16" "2016-04-17" "2016-04-18" "2016-04-19" "2016-04-20"
#>  [831] "2016-04-21" "2016-04-22" "2016-04-23" "2016-04-24" "2016-04-25"
#>  [836] "2016-04-26" "2016-04-27" "2016-04-28" "2016-04-29" "2016-04-30"
#>  [841] "2016-05-01" "2016-05-02" "2016-05-03" "2016-05-04" "2016-05-05"
#>  [846] "2016-05-06" "2016-05-07" "2016-05-08" "2016-05-09" "2016-05-10"
#>  [851] "2016-05-11" "2016-05-12" "2016-05-13" "2016-05-14" "2016-05-15"
#>  [856] "2016-05-16" "2016-05-17" "2016-05-18" "2016-05-19" "2016-05-20"
#>  [861] "2016-05-21" "2016-05-22" "2016-05-23" "2016-05-24" "2016-05-25"
#>  [866] "2016-05-26" "2016-05-27" "2016-05-28" "2016-05-29" "2016-05-30"
#>  [871] "2016-06-01" "2016-06-02" "2016-06-03" "2016-06-04" "2016-06-05"
#>  [876] "2016-06-06" "2016-06-07" "2016-06-08" "2016-06-09" "2016-06-10"
#>  [881] "2016-06-11" "2016-06-12" "2016-06-13" "2016-06-14" "2016-06-15"
#>  [886] "2016-06-16" "2016-06-17" "2016-06-18" "2016-06-19" "2016-06-20"
#>  [891] "2016-06-21" "2016-06-22" "2016-06-23" "2016-06-24" "2016-06-25"
#>  [896] "2016-06-26" "2016-06-27" "2016-06-28" "2016-06-29" "2016-06-30"
#>  [901] "2016-07-01" "2016-07-02" "2016-07-03" "2016-07-04" "2016-07-05"
#>  [906] "2016-07-06" "2016-07-07" "2016-07-08" "2016-07-09" "2016-07-10"
#>  [911] "2016-07-11" "2016-07-12" "2016-07-13" "2016-07-14" "2016-07-15"
#>  [916] "2016-07-16" "2016-07-17" "2016-07-18" "2016-07-19" "2016-07-20"
#>  [921] "2016-07-21" "2016-07-22" "2016-07-23" "2016-07-24" "2016-07-25"
#>  [926] "2016-07-26" "2016-07-27" "2016-07-28" "2016-07-29" "2016-07-30"
#>  [931] "2016-08-01" "2016-08-02" "2016-08-03" "2016-08-04" "2016-08-05"
#>  [936] "2016-08-06" "2016-08-07" "2016-08-08" "2016-08-09" "2016-08-10"
#>  [941] "2016-08-11" "2016-08-12" "2016-08-13" "2016-08-14" "2016-08-15"
#>  [946] "2016-08-16" "2016-08-17" "2016-08-18" "2016-08-19" "2016-08-20"
#>  [951] "2016-08-21" "2016-08-22" "2016-08-23" "2016-08-24" "2016-08-25"
#>  [956] "2016-08-26" "2016-08-27" "2016-08-28" "2016-08-29" "2016-08-30"
#>  [961] "2016-09-01" "2016-09-02" "2016-09-03" "2016-09-04" "2016-09-05"
#>  [966] "2016-09-06" "2016-09-07" "2016-09-08" "2016-09-09" "2016-09-10"
#>  [971] "2016-09-11" "2016-09-12" "2016-09-13" "2016-09-14" "2016-09-15"
#>  [976] "2016-09-16" "2016-09-17" "2016-09-18" "2016-09-19" "2016-09-20"
#>  [981] "2016-09-21" "2016-09-22" "2016-09-23" "2016-09-24" "2016-09-25"
#>  [986] "2016-09-26" "2016-09-27" "2016-09-28" "2016-09-29" "2016-09-30"
#>  [991] "2016-10-01" "2016-10-02" "2016-10-03" "2016-10-04" "2016-10-05"
#>  [996] "2016-10-06" "2016-10-07" "2016-10-08" "2016-10-09" "2016-10-10"
#>  [ reached 'max' / getOption("max.print") -- omitted 2240 entries ]