Saving Climate Data From ERA5 and CHIRTSdaily

Saving Climate Data From ERA5 and CHIRTSdaily

Usage

save_climate_data(
  lon,
  lat,
  years,
  path_to_data,
  rain = TRUE,
  temp = TRUE,
  temp_file_name = NULL,
  rain_file_name = NULL
)

Arguments

lon

longitude coordinate

lat

latitude coordinate

years

vector containing each year for data to be downloaded

path_to_data

folder where data is to be saved

rain

boolean for if CHIRPS data should be downloaded and saved

temp

boolean for if ERA5 data should be downloaded and saved

temp_file_name

character file name for stored temperature data (ERA5)

rain_file_name

character file name for stored rainfall data (CHIRPS)

Value

nothing

Examples

years <- 2014:2023
lon <- 17.9
lat <- 8.3
path_to_data <- "C:/Users/putnni/Documents/r-packages/data/"
save_climate_data(lon, lat, years, path_to_data)
#> Requesting data to the cds service with username ecmwfr
#> Superclass ecmwfr_service has cloneable=FALSE, but subclass ecmwfr_ds has cloneable=TRUE. A subclass cannot be cloneable when its superclass is not cloneable, so cloning will be disabled for ecmwfr_ds.
#> - staging data transfer at url endpoint or request id:
#>   403ae997-6c07-4aaf-8e59-41aa9951cdb2
#>   on server: https://cds.climate.copernicus.eu/api
#> - timeout set to 1.0 hours
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#> \ polling server ... 
#> | polling server ... 
#> / polling server ... 
#> - polling server ... 
#>    file ready to download...           
#> 
#> Downloading file
#> Error in curl::curl_fetch_disk(url, x$path, handle = handle): Failed to open file C:\Users\putnni\Documents\r-packages\data\ecmwfr_6aec985533a.