Convert Climate Data to 360-Day Years

This function converts climate data to a 360-day calendar, with each month forced to 30 days. Data is either removed or repeated to ensure the adjustment, useful for simplifying climate simulations.

Usage

climate_to_30_day_months(clim_df, start_year = 2014, end_year = 2022)

Arguments

clim_df

A data frame of climate data, which must include columns for date, month, day, anom, temp, and rollrain.

start_year

The starting year of the conversion (default is 2014).

end_year

The ending year of the conversion (default is 2022).

Value

A data frame with 360-day years, where each month has exactly 30 days.

Examples

climate_to_30_day_months(clim_df, start_year = 2014, end_year = 2022)
#> Error: object 'clim_df' not found