Convert Climate Data to 360-Day Years

This function converts climate data to a 360-day calendar, with each month forced to 30 days. Data is either removed or repeated to ensure the adjustment, useful for simplifying climate simulations.

Usage

climate_to_30_day_months(clim_df, start_year = 2014, end_year = 2022)

Arguments

clim_df

A data frame of climate data, which must include columns for date, month, day, anom, temp, and CumulativeRainfall.

start_year

The starting year of the conversion (default is 2014).

end_year

The ending year of the conversion (default is 2022).

Value

A data frame with 360-day years, where each month has exactly 30 days.

Examples

library(lubridate)
# Create example climate data for 1 year
example_dates <- seq(as.Date("2014-01-01"), as.Date("2014-12-31"), by = "day")
clim_df <- data.frame(
  date = example_dates,
  month = month(example_dates),
  day = day(example_dates),
  anom = rnorm(length(example_dates)),
  temp = runif(length(example_dates), min = 20, max = 35),
  CumulativeRainfall = cumsum(runif(length(example_dates), 0, 5))
)
# Run conversion
climate_to_30_day_months(clim_df, start_year = 2014, end_year = 2014)
#>           date         anom     temp     cumrain
#> 1   2014-01-01 -0.005571287 32.96662   0.6002071
#> 2   2014-01-02  0.621552721 21.15635   1.4515254
#> 3   2014-01-03  1.148411606 26.46505   4.0591699
#> 4   2014-01-04 -1.821817661 34.60434   8.8766400
#> 5   2014-01-05 -0.247325302 28.56436  13.3160026
#> 6   2014-01-06 -0.244199607 31.31610  16.4982356
#> 7   2014-01-07 -0.282705449 23.63681  18.1451368
#> 8   2014-01-08 -0.553699384 29.40737  21.1562489
#> 9   2014-01-09  0.628982042 22.52622  25.0170566
#> 10  2014-01-10  2.065024895 26.67184  26.0677330
#> 11  2014-01-11 -1.630989402 22.47013  28.9892749
#> 12  2014-01-12  0.512426950 21.77415  31.6996539
#> 13  2014-01-13 -1.863011492 25.51197  32.1616371
#> 14  2014-01-14 -0.522012515 34.19838  34.0083413
#> 15  2014-01-15 -0.052601910 33.13116  38.7704884
#> 16  2014-01-16  0.542996343 23.90088  43.5189346
#> 17  2014-01-17 -0.914074827 29.85948  45.7678187
#> 18  2014-01-18  0.468154420 21.82797  47.0669099
#> 19  2014-01-19  0.362951256 24.78272  47.1614408
#> 20  2014-01-20 -1.304543545 29.34592  50.7684660
#> 21  2014-01-21  0.737776321 26.70496  55.5325142
#> 22  2014-01-22  1.888504929 30.62440  57.7336308
#> 23  2014-01-23 -0.097445104 33.35043  60.2536927
#> 24  2014-01-24 -0.935847354 21.56460  60.4714777
#> 25  2014-01-25 -0.015950311 29.45618  65.0838499
#> 26  2014-01-26 -0.826788954 32.62338  69.7027977
#> 27  2014-01-27 -1.512399651 25.46394  72.7310363
#> 28  2014-01-28  0.935363190 29.34384  72.7514462
#> 29  2014-01-29  0.176488611 26.91121  76.0357284
#> 30  2014-01-30  0.243685465 22.45118  76.7903731
#> 31  2014-02-01  0.112038083 22.14900  78.9843681
#> 32  2014-02-02 -0.133997013 29.75448  82.2922379
#> 33  2014-02-03 -1.910087468 22.32942  83.5011820
#> 34  2014-02-04 -0.279237242 27.62465  83.8589307
#> 35  2014-02-05 -0.313445978 21.28806  87.3822904
#> 36  2014-02-06  1.067307879 27.54540  88.9592749
#> 37  2014-02-07  0.070034850 20.57858  90.8386509
#> 38  2014-02-08 -0.639123324 32.36049  94.9141396
#> 39  2014-02-09 -0.049964899 24.28978  95.9169394
#> 40  2014-02-10 -0.251483443 23.70126  97.4265904
#> 41  2014-02-11  0.444797116 25.96990  97.8651916
#> 42  2014-02-12  2.755417575 29.48139  98.7817922
#> 43  2014-02-13  0.046531380 34.22868  99.8158016
#> 44  2014-02-14  0.577709069 25.10179 104.0905991
#> 45  2014-02-15  0.118194874 29.41049 105.4051486
#> 46  2014-02-16 -1.911720491 32.37373 110.1318200
#> 47  2014-02-17  0.862086482 22.17673 114.0954973
#> 48  2014-02-18 -0.243236740 34.50709 118.2534068
#> 49  2014-02-19 -0.206087195 21.29058 121.3482929
#> 50  2014-02-20  0.019177592 23.60692 123.5669610
#> 51  2014-02-21  0.029560754 29.58326 128.3900891
#> 52  2014-02-22  0.549827542 27.55093 129.2010826
#> 53  2014-02-23 -2.274114857 23.75785 130.1293516
#> 54  2014-02-24  2.682557184 34.68551 133.1948587
#> 55  2014-02-25 -0.361221255 29.84869 135.6022811
#> 56  2014-02-26  0.213355750 21.34773 137.5829705
#> 57  2014-02-27  1.074345882 29.45814 139.1767526
#> 58  2014-02-28 -0.665088249 33.15151 143.3686276
#> 59  2014-02-29 -0.665088249 33.15151 143.3686276
#> 60  2014-02-30 -0.665088249 33.15151 143.3686276
#> 61  2014-03-01  1.113952419 33.95422 144.8646453
#> 62  2014-03-02 -0.245896412 26.29229 147.9449544
#> 63  2014-03-03 -1.177563309 27.43340 150.2554736
#> 64  2014-03-04 -0.975850616 25.29057 154.9515034
#> 65  2014-03-05  1.065057320 23.73102 154.9829603
#> 66  2014-03-06  0.131670635 34.38092 156.7300977
#> 67  2014-03-07  0.488628809 25.39940 159.5614547
#> 68  2014-03-08 -1.699450568 22.98479 162.2062400
#> 69  2014-03-09 -1.470736306 30.89428 165.3913170
#> 70  2014-03-10  0.284150344 22.52199 167.1452964
#> 71  2014-03-11  1.337320413 23.39722 171.7661524
#> 72  2014-03-12  0.236696283 32.68310 174.1363585
#> 73  2014-03-13  1.318293384 25.18780 178.0270291
#> 74  2014-03-14  0.523909788 21.00980 182.0207118
#> 75  2014-03-15  0.606748047 20.40529 186.4132113
#> 76  2014-03-16 -0.109935672 21.77067 190.9624306
#> 77  2014-03-17  0.172181715 20.20543 193.1901194
#> 78  2014-03-18 -0.090327287 31.03594 194.6379160
#> 79  2014-03-19  1.924343341 33.85941 196.1544697
#> 80  2014-03-20  1.298392759 34.73286 200.7495915
#> 81  2014-03-21  0.748791268 26.07905 205.4281534
#> 82  2014-03-22  0.556224329 24.04542 208.7863951
#> 83  2014-03-23 -0.548257264 27.32433 210.6427449
#> 84  2014-03-24  1.110534893 20.76594 213.4784021
#> 85  2014-03-25 -2.612334333 20.03715 215.8713943
#> 86  2014-03-26 -0.155693776 27.43752 218.7101382
#> 87  2014-03-27  0.433889790 23.65687 220.1428324
#> 88  2014-03-28 -0.381951112 23.83624 223.9408281
#> 89  2014-03-29  0.424187575 28.28478 225.0831569
#> 90  2014-03-30  1.063101996 23.77366 227.2156839
#> 91  2014-04-01 -0.038102895 24.74173 233.0481413
#> 92  2014-04-02  0.486148920 33.86781 234.4854262
#> 93  2014-04-03  1.672882611 33.59738 238.3507275
#> 94  2014-04-04 -0.354361164 21.37955 241.2235390
#> 95  2014-04-05  0.946347886 31.06693 241.3325178
#> 96  2014-04-06  1.316826356 28.54611 245.1976502
#> 97  2014-04-07 -0.296640025 30.44733 245.4932948
#> 98  2014-04-08 -0.387213575 30.51151 245.6315363
#> 99  2014-04-09 -0.785432656 20.59911 249.3357693
#> 100 2014-04-10 -1.056736867 26.47284 252.3023735
#> 101 2014-04-11 -0.795541430 32.21347 255.7357571
#> 102 2014-04-12 -1.756275428 28.49879 259.4123083
#> 103 2014-04-13 -0.690537897 28.82371 261.7178343
#> 104 2014-04-14 -0.558541994 33.20464 265.0278307
#> 105 2014-04-15 -0.536663326 30.78977 265.0326204
#> 106 2014-04-16  0.227127133 34.04643 269.9124481
#> 107 2014-04-17  0.978454920 26.44147 271.0132653
#> 108 2014-04-18 -0.208882651 24.29281 272.5449045
#> 109 2014-04-19 -1.399410460 31.55965 276.2027377
#> 110 2014-04-20  0.258537288 29.24748 276.8702688
#> 111 2014-04-21 -0.441799453 22.38912 277.3694491
#> 112 2014-04-22  0.568599861 26.02397 278.2263749
#> 113 2014-04-23  2.126850459 20.02470 281.3122829
#> 114 2014-04-24  0.424858441 26.26949 284.7285994
#> 115 2014-04-25 -1.684281532 31.45902 288.6814697
#> 116 2014-04-26  0.249401784 31.91717 289.7477104
#> 117 2014-04-27  1.072838252 23.63881 293.7077648
#> 118 2014-04-28  2.039369263 31.38756 296.6709369
#> 119 2014-04-29  0.449453778 20.43656 299.7022549
#> 120 2014-04-30  1.391814046 26.56011 300.2560976
#> 121 2014-05-01  0.426566547 27.95614 301.2341408
#> 122 2014-05-02  0.107583992 26.44747 301.2694947
#> 123 2014-05-03  0.022294733 32.13945 305.9738476
#> 124 2014-05-04  0.603611011 33.83847 306.3148944
#> 125 2014-05-05 -0.262650573 30.12535 306.9995403
#> 126 2014-05-06 -0.528264082 21.40952 308.6636254
#> 127 2014-05-07  0.192149422 22.96195 311.1120669
#> 128 2014-05-08 -1.146199669 27.13318 311.9768396
#> 129 2014-05-09  0.846184665 30.71557 312.3015358
#> 130 2014-05-10  0.081719629 28.07119 314.5598707
#> 131 2014-05-11 -1.305117010 33.13075 316.4568825
#> 132 2014-05-12 -0.944912060 24.40378 321.0033743
#> 133 2014-05-13  0.454341594 23.37148 321.1531066
#> 134 2014-05-14 -0.855202501 30.36377 322.1270884
#> 135 2014-05-15 -0.286895219 24.68494 325.5244853
#> 136 2014-05-16  0.894961626 27.24947 329.2618168
#> 137 2014-05-17  0.067304440 21.82695 332.5095054
#> 138 2014-05-18 -0.162676337 30.35011 332.8974088
#> 139 2014-05-19 -0.827310169 24.73597 335.8956770
#> 140 2014-05-20  1.876505621 32.24282 340.2907311
#> 141 2014-05-21  0.766440199 28.18873 343.9509169
#> 142 2014-05-22  0.979956696 34.20976 344.3511823
#> 143 2014-05-23  1.321780992 20.57790 348.5149314
#> 144 2014-05-24 -1.119710829 29.71466 352.8384599
#> 145 2014-05-25  0.514599819 21.19429 356.4032847
#> 146 2014-05-26 -1.509099836 26.49396 360.1903203
#> 147 2014-05-27  1.532741480 31.41340 360.8024899
#> 148 2014-05-28  0.429147371 24.40638 361.7709793
#> 149 2014-05-29  0.122103414 21.60998 363.3418707
#> 150 2014-05-30 -1.138012401 21.09492 364.0295356
#> 151 2014-06-01  1.052538537 28.57341 370.6946600
#> 152 2014-06-02  0.677683644 30.16982 373.0458736
#> 153 2014-06-03  0.038499547 30.36079 377.0011005
#> 154 2014-06-04 -0.356381187 25.60116 378.5660168
#> 155 2014-06-05  0.782844102 21.46967 380.1098394
#> 156 2014-06-06  0.804411616 21.48515 381.8356161
#> 157 2014-06-07 -1.900060823 31.62202 385.9778327
#> 158 2014-06-08  0.935784286 22.27243 387.3478101
#> 159 2014-06-09 -0.309051503 25.91910 389.5019906
#> 160 2014-06-10  0.263066677 34.02498 391.5864798
#> 161 2014-06-11 -1.790591856 31.09384 392.2580408
#> 162 2014-06-12 -0.788258845 29.53502 395.6897478
#> 163 2014-06-13 -1.133021669 25.07330 396.5501383
#> 164 2014-06-14  0.363652568 34.43624 401.1676056
#> 165 2014-06-15 -0.285887914 21.67685 404.2363235
#> 166 2014-06-16  0.517669134 29.75728 405.8144708
#> 167 2014-06-17 -0.102908670 34.75267 409.8631362
#> 168 2014-06-18 -0.974069593 22.68691 414.3750067
#> 169 2014-06-19  1.270672301 22.06956 415.0014957
#> 170 2014-06-20  0.960864787 20.84838 417.8292817
#> 171 2014-06-21  0.768721370 28.19551 418.4674872
#> 172 2014-06-22  1.035930771 26.99757 421.6054753
#> 173 2014-06-23 -0.473887074 33.44568 421.9820951
#> 174 2014-06-24 -1.275334875 20.27797 423.4093517
#> 175 2014-06-25 -0.305620674 32.76461 424.9623638
#> 176 2014-06-26  2.211769487 27.52135 425.6960016
#> 177 2014-06-27 -1.041668381 22.49243 429.6948542
#> 178 2014-06-28 -1.146523850 25.41509 430.5548346
#> 179 2014-06-29 -1.675327303 29.78000 434.4832537
#> 180 2014-06-30  1.525938655 33.11265 436.8334325
#> 181 2014-07-01  0.554185515 30.92458 438.9982211
#> 182 2014-07-02  1.993110265 26.18709 441.8708854
#> 183 2014-07-03 -0.154120740 22.38844 445.5207352
#> 184 2014-07-04  2.564408338 32.68555 449.9916709
#> 185 2014-07-05  1.061999145 33.26554 451.0932028
#> 186 2014-07-06  1.142694878 25.93997 452.5511849
#> 187 2014-07-07  1.123838843 20.57211 456.6156787
#> 188 2014-07-08 -0.397001493 34.27094 456.8336903
#> 189 2014-07-09 -0.823261151 31.75591 459.3600487
#> 190 2014-07-10 -0.578884625 23.29160 460.5683281
#> 191 2014-07-11  1.763789378 24.96398 463.5140402
#> 192 2014-07-12  0.132992146 22.66639 468.3697430
#> 193 2014-07-13  0.376499328 25.43774 469.5333480
#> 194 2014-07-14  1.138707653 25.87180 473.8704381
#> 195 2014-07-15  1.241263075 33.55816 474.4242468
#> 196 2014-07-16  0.612090945 24.14685 478.4317057
#> 197 2014-07-17 -0.429380087 22.09039 481.7213280
#> 198 2014-07-18  1.360461327 26.77004 484.9815995
#> 199 2014-07-19 -0.070857431 21.65495 486.1132127
#> 200 2014-07-20 -0.272153684 33.21435 490.3831108
#> 201 2014-07-21 -2.446680029 28.00764 492.5611460
#> 202 2014-07-22  0.065486641 31.08876 494.0075547
#> 203 2014-07-23 -1.098508902 30.79284 497.1954781
#> 204 2014-07-24 -0.633178176 23.96723 498.5181754
#> 205 2014-07-25 -2.063654451 30.10785 500.4495173
#> 206 2014-07-26  2.648932029 23.14432 503.5454871
#> 207 2014-07-27 -1.153398386 30.35608 506.2538775
#> 208 2014-07-28 -0.340637876 22.92139 508.6780325
#> 209 2014-07-29  0.786362576 31.73231 512.4995071
#> 210 2014-07-30 -1.270513110 20.22548 517.3046418
#> 211 2014-08-01  0.075105900 21.26654 522.3105656
#> 212 2014-08-02  0.558514422 20.51078 522.9813933
#> 213 2014-08-03  0.415406399 24.66439 523.2233845
#> 214 2014-08-04 -1.452299769 31.77065 526.7414198
#> 215 2014-08-05  0.941206122 24.20601 527.4774888
#> 216 2014-08-06 -0.338935872 22.47926 531.3232654
#> 217 2014-08-07 -0.075574247 26.50392 532.2050160
#> 218 2014-08-08  0.040204392 32.21068 532.7588110
#> 219 2014-08-09  0.124301066 30.29771 537.4505230
#> 220 2014-08-10 -0.998432551 31.65959 541.6828787
#> 221 2014-08-11  1.233390065 21.36801 544.5385966
#> 222 2014-08-12  0.340424488 31.13936 547.9340637
#> 223 2014-08-13 -0.472702482 25.95011 548.3806749
#> 224 2014-08-14  0.708753061 24.44042 552.3075455
#> 225 2014-08-15 -1.528958715 31.14386 553.4423328
#> 226 2014-08-16  0.237425345 26.99764 555.6832251
#> 227 2014-08-17 -1.312814246 31.59542 556.4893633
#> 228 2014-08-18  0.747028587 20.33176 557.3699217
#> 229 2014-08-19 -1.562518435 32.22535 558.3610623
#> 230 2014-08-20  0.071053360 24.55750 560.1491189
#> 231 2014-08-21 -0.639534770 33.97694 561.0557849
#> 232 2014-08-22 -0.845195739 28.10647 563.8615092
#> 233 2014-08-23  0.675244698 23.82228 567.1569243
#> 234 2014-08-24  1.153375794 27.16381 570.4581913
#> 235 2014-08-25 -1.686504742 30.56696 570.4701952
#> 236 2014-08-26 -0.902814949 20.11258 575.4374254
#> 237 2014-08-27  1.317633698 31.58613 578.5748716
#> 238 2014-08-28  1.100189745 23.98690 578.6480233
#> 239 2014-08-29  1.203767839 34.56561 579.6739143
#> 240 2014-08-30 -1.431270777 22.85788 582.9892977
#> 241 2014-09-01  0.003125940 29.18565 587.1097586
#> 242 2014-09-02 -0.077886824 27.88597 590.5565906
#> 243 2014-09-03  0.441428226 29.51780 591.8508945
#> 244 2014-09-04  0.128922896 25.93172 596.1069197
#> 245 2014-09-05 -0.830214260 22.62302 598.5069431
#> 246 2014-09-06 -0.503592910 24.90970 601.1932751
#> 247 2014-09-07 -1.193641182 29.22130 604.5014746
#> 248 2014-09-08 -0.751723323 21.18750 606.0057471
#> 249 2014-09-09  1.455841403 27.78591 607.3232547
#> 250 2014-09-10 -0.828603533 24.59422 608.8484853
#> 251 2014-09-11  0.289774460 30.07093 613.2232767
#> 252 2014-09-12 -0.480053484 25.90508 616.8154627
#> 253 2014-09-13 -0.604829354 31.21772 618.8500612
#> 254 2014-09-14  1.460110180 22.08390 618.9468586
#> 255 2014-09-15  0.149679354 25.66015 619.1977604
#> 256 2014-09-16 -1.433321100 29.62075 622.7090155
#> 257 2014-09-17 -0.010303319 34.91292 623.0408669
#> 258 2014-09-18 -0.212236035 25.52839 623.2215287
#> 259 2014-09-19 -0.906340179 25.06023 623.3222972
#> 260 2014-09-20 -2.102152479 33.70528 625.1862027
#> 261 2014-09-21  1.893360464 33.64131 626.3145297
#> 262 2014-09-22 -0.968125837 23.57570 628.7741202
#> 263 2014-09-23 -0.102603036 26.17520 631.0073460
#> 264 2014-09-24  0.239959572 31.19108 634.4583705
#> 265 2014-09-25  0.060898893 30.85176 638.6720704
#> 266 2014-09-26 -2.177576028 29.21786 640.5512387
#> 267 2014-09-27 -0.117860143 27.10971 645.4756872
#> 268 2014-09-28  0.112294787 28.67692 650.2378188
#> 269 2014-09-29  0.007886198 23.62391 655.1873114
#> 270 2014-09-30  1.877743872 21.71568 657.4029531
#> 271 2014-10-01  2.158756554 24.47517 658.1717724
#> 272 2014-10-02  0.709714522 27.26179 662.8954090
#> 273 2014-10-03  0.766983379 28.51109 665.4927411
#> 274 2014-10-04 -0.308211421 31.30930 667.7709141
#> 275 2014-10-05  1.012001849 28.18138 668.8713651
#> 276 2014-10-06 -0.919051597 31.39517 669.5672770
#> 277 2014-10-07  0.563380077 23.19451 670.6480664
#> 278 2014-10-08  0.322482749 32.58898 672.6764598
#> 279 2014-10-09  0.366674363 26.10046 673.8451327
#> 280 2014-10-10  1.129835153 33.10610 673.9704645
#> 281 2014-10-11 -0.941498076 33.23235 677.7321743
#> 282 2014-10-12  0.217837643 21.61534 681.3925370
#> 283 2014-10-13  1.415412293 28.61817 683.4380709
#> 284 2014-10-14 -0.383733048 34.94050 686.5923599
#> 285 2014-10-15 -0.174086374 32.75830 686.9195849
#> 286 2014-10-16 -0.221744517 26.56483 687.6630456
#> 287 2014-10-17 -1.009528722 24.89645 689.0615544
#> 288 2014-10-18  0.480725266 25.03953 693.3148385
#> 289 2014-10-19  1.604407328 23.58911 693.4225424
#> 290 2014-10-20 -1.515024529 26.32057 694.6866918
#> 291 2014-10-21 -1.416023914 27.68205 697.9702493
#> 292 2014-10-22  0.876777327 22.78170 701.9159963
#> 293 2014-10-23  0.624132413 24.89601 705.4602246
#> 294 2014-10-24  2.112277288 32.25610 705.7831895
#> 295 2014-10-25 -0.356124416 31.23107 708.3397463
#> 296 2014-10-26 -1.064464209 27.97613 709.3443880
#> 297 2014-10-27  1.077116538 34.42357 713.4149660
#> 298 2014-10-28  1.181575567 21.56070 716.6838664
#> 299 2014-10-29  0.198392095 31.05685 718.4501263
#> 300 2014-10-30 -0.400405249 32.71220 721.4331007
#> 301 2014-11-01  1.974156748 23.48686 730.0494653
#> 302 2014-11-02  1.884662324 33.74344 732.5397417
#> 303 2014-11-03 -1.588620547 26.36896 733.6912155
#> 304 2014-11-04 -0.539923164 22.89415 735.1974871
#> 305 2014-11-05 -1.169461464 30.52326 738.7998205
#> 306 2014-11-06  0.559105989 34.88871 740.1068476
#> 307 2014-11-07 -1.819347247 30.63369 740.6453293
#> 308 2014-11-08  0.393343972 27.76228 741.9574429
#> 309 2014-11-09  0.042134106 30.63098 742.8909738
#> 310 2014-11-10  1.179664177 25.62711 745.6031822
#> 311 2014-11-11 -0.256921176 23.82161 745.9307745
#> 312 2014-11-12 -1.056336098 34.94619 750.3327002
#> 313 2014-11-13  0.198777205 27.84343 751.7287629
#> 314 2014-11-14  0.650533552 30.73714 755.0922813
#> 315 2014-11-15  0.343913337 20.03739 755.3929356
#> 316 2014-11-16  1.477532312 24.39453 760.3353634
#> 317 2014-11-17  0.072025698 22.71224 761.7467405
#> 318 2014-11-18  2.126444534 26.11391 764.5173024
#> 319 2014-11-19 -1.476196906 23.25494 768.9677224
#> 320 2014-11-20  0.407888500 21.07014 769.0817327
#> 321 2014-11-21  1.393977798 23.79967 772.4701749
#> 322 2014-11-22  0.360278296 29.78719 774.1554739
#> 323 2014-11-23  0.654550251 31.01666 774.8614544
#> 324 2014-11-24  1.052155422 32.26270 776.6242743
#> 325 2014-11-25 -1.979555125 24.58986 778.6314986
#> 326 2014-11-26  1.208385605 33.27871 783.0760704
#> 327 2014-11-27 -0.169280084 29.09838 784.6391413
#> 328 2014-11-28  0.295029753 33.21529 784.9474149
#> 329 2014-11-29  1.266340587 30.19881 788.3448462
#> 330 2014-11-30 -1.135343257 22.14415 790.2356590
#> 331 2014-12-01 -1.131053798 31.48198 793.1775108
#> 332 2014-12-02  0.109993384 29.40450 795.6374034
#> 333 2014-12-03  0.852905410 30.95508 797.0662882
#> 334 2014-12-04 -0.234337862 26.35557 797.1083861
#> 335 2014-12-05  2.086688567 24.02496 799.8622576
#> 336 2014-12-06 -0.110919371 31.53572 803.9284122
#> 337 2014-12-07 -1.392847056 28.81020 807.1462851
#> 338 2014-12-08 -1.142290768 33.47852 807.4915951
#> 339 2014-12-09  1.704608737 33.05487 811.1351765
#> 340 2014-12-10 -0.080073634 25.98959 814.4957325
#> 341 2014-12-11 -0.437281240 34.47679 818.5947374
#> 342 2014-12-12 -0.119215094 27.43987 823.1678452
#> 343 2014-12-13  0.786462865 26.99896 826.6176986
#> 344 2014-12-14 -0.578945246 21.61617 829.0775876
#> 345 2014-12-15 -0.145426885 33.71447 832.3122309
#> 346 2014-12-16  0.526457991 22.15705 834.5948057
#> 347 2014-12-17  1.733578110 23.97781 835.2220388
#> 348 2014-12-18  1.448657220 22.22695 838.2438636
#> 349 2014-12-19  1.518193149 26.21373 842.3779973
#> 350 2014-12-20 -0.384007254 21.63425 845.4748886
#> 351 2014-12-21  1.827125177 23.70617 849.6801571
#> 352 2014-12-22 -0.551491750 28.85965 851.8163902
#> 353 2014-12-23 -0.865753541 24.92401 855.6628237
#> 354 2014-12-24 -0.343831481 30.99456 859.2235276
#> 355 2014-12-25  1.062876458 30.91858 859.5403637
#> 356 2014-12-26  0.813058204 28.18585 861.6703899
#> 357 2014-12-27  1.803483361 31.00619 862.1117781
#> 358 2014-12-28 -0.105068694 34.00038 863.1533622
#> 359 2014-12-29  0.982453362 23.65089 864.2946665
#> 360 2014-12-30 -1.713302622 24.24439 864.5468773