Generate Dates for a 360-Day Year Calendar

This function generates a sequence of dates assuming a 360-day year, with each month consisting of 30 days. Useful for simplified models where leap years and 31-day months are ignored.

Usage

generate_360_day_dates(start_year, end_year)

Arguments

start_year

The starting year of the date sequence.

end_year

The ending year of the date sequence.

Value

A vector of dates formatted as "YYYY-MM-DD" for each day within the given range, assuming a 360-day year.

Examples

generate_360_day_dates(2014, 2022)
#>    [1] "2014-01-01" "2014-01-02" "2014-01-03" "2014-01-04" "2014-01-05"
#>    [6] "2014-01-06" "2014-01-07" "2014-01-08" "2014-01-09" "2014-01-10"
#>   [11] "2014-01-11" "2014-01-12" "2014-01-13" "2014-01-14" "2014-01-15"
#>   [16] "2014-01-16" "2014-01-17" "2014-01-18" "2014-01-19" "2014-01-20"
#>   [21] "2014-01-21" "2014-01-22" "2014-01-23" "2014-01-24" "2014-01-25"
#>   [26] "2014-01-26" "2014-01-27" "2014-01-28" "2014-01-29" "2014-01-30"
#>   [31] "2014-02-01" "2014-02-02" "2014-02-03" "2014-02-04" "2014-02-05"
#>   [36] "2014-02-06" "2014-02-07" "2014-02-08" "2014-02-09" "2014-02-10"
#>   [41] "2014-02-11" "2014-02-12" "2014-02-13" "2014-02-14" "2014-02-15"
#>   [46] "2014-02-16" "2014-02-17" "2014-02-18" "2014-02-19" "2014-02-20"
#>   [51] "2014-02-21" "2014-02-22" "2014-02-23" "2014-02-24" "2014-02-25"
#>   [56] "2014-02-26" "2014-02-27" "2014-02-28" "2014-02-29" "2014-02-30"
#>   [61] "2014-03-01" "2014-03-02" "2014-03-03" "2014-03-04" "2014-03-05"
#>   [66] "2014-03-06" "2014-03-07" "2014-03-08" "2014-03-09" "2014-03-10"
#>   [71] "2014-03-11" "2014-03-12" "2014-03-13" "2014-03-14" "2014-03-15"
#>   [76] "2014-03-16" "2014-03-17" "2014-03-18" "2014-03-19" "2014-03-20"
#>   [81] "2014-03-21" "2014-03-22" "2014-03-23" "2014-03-24" "2014-03-25"
#>   [86] "2014-03-26" "2014-03-27" "2014-03-28" "2014-03-29" "2014-03-30"
#>   [91] "2014-04-01" "2014-04-02" "2014-04-03" "2014-04-04" "2014-04-05"
#>   [96] "2014-04-06" "2014-04-07" "2014-04-08" "2014-04-09" "2014-04-10"
#>  [101] "2014-04-11" "2014-04-12" "2014-04-13" "2014-04-14" "2014-04-15"
#>  [106] "2014-04-16" "2014-04-17" "2014-04-18" "2014-04-19" "2014-04-20"
#>  [111] "2014-04-21" "2014-04-22" "2014-04-23" "2014-04-24" "2014-04-25"
#>  [116] "2014-04-26" "2014-04-27" "2014-04-28" "2014-04-29" "2014-04-30"
#>  [121] "2014-05-01" "2014-05-02" "2014-05-03" "2014-05-04" "2014-05-05"
#>  [126] "2014-05-06" "2014-05-07" "2014-05-08" "2014-05-09" "2014-05-10"
#>  [131] "2014-05-11" "2014-05-12" "2014-05-13" "2014-05-14" "2014-05-15"
#>  [136] "2014-05-16" "2014-05-17" "2014-05-18" "2014-05-19" "2014-05-20"
#>  [141] "2014-05-21" "2014-05-22" "2014-05-23" "2014-05-24" "2014-05-25"
#>  [146] "2014-05-26" "2014-05-27" "2014-05-28" "2014-05-29" "2014-05-30"
#>  [151] "2014-06-01" "2014-06-02" "2014-06-03" "2014-06-04" "2014-06-05"
#>  [156] "2014-06-06" "2014-06-07" "2014-06-08" "2014-06-09" "2014-06-10"
#>  [161] "2014-06-11" "2014-06-12" "2014-06-13" "2014-06-14" "2014-06-15"
#>  [166] "2014-06-16" "2014-06-17" "2014-06-18" "2014-06-19" "2014-06-20"
#>  [171] "2014-06-21" "2014-06-22" "2014-06-23" "2014-06-24" "2014-06-25"
#>  [176] "2014-06-26" "2014-06-27" "2014-06-28" "2014-06-29" "2014-06-30"
#>  [181] "2014-07-01" "2014-07-02" "2014-07-03" "2014-07-04" "2014-07-05"
#>  [186] "2014-07-06" "2014-07-07" "2014-07-08" "2014-07-09" "2014-07-10"
#>  [191] "2014-07-11" "2014-07-12" "2014-07-13" "2014-07-14" "2014-07-15"
#>  [196] "2014-07-16" "2014-07-17" "2014-07-18" "2014-07-19" "2014-07-20"
#>  [201] "2014-07-21" "2014-07-22" "2014-07-23" "2014-07-24" "2014-07-25"
#>  [206] "2014-07-26" "2014-07-27" "2014-07-28" "2014-07-29" "2014-07-30"
#>  [211] "2014-08-01" "2014-08-02" "2014-08-03" "2014-08-04" "2014-08-05"
#>  [216] "2014-08-06" "2014-08-07" "2014-08-08" "2014-08-09" "2014-08-10"
#>  [221] "2014-08-11" "2014-08-12" "2014-08-13" "2014-08-14" "2014-08-15"
#>  [226] "2014-08-16" "2014-08-17" "2014-08-18" "2014-08-19" "2014-08-20"
#>  [231] "2014-08-21" "2014-08-22" "2014-08-23" "2014-08-24" "2014-08-25"
#>  [236] "2014-08-26" "2014-08-27" "2014-08-28" "2014-08-29" "2014-08-30"
#>  [241] "2014-09-01" "2014-09-02" "2014-09-03" "2014-09-04" "2014-09-05"
#>  [246] "2014-09-06" "2014-09-07" "2014-09-08" "2014-09-09" "2014-09-10"
#>  [251] "2014-09-11" "2014-09-12" "2014-09-13" "2014-09-14" "2014-09-15"
#>  [256] "2014-09-16" "2014-09-17" "2014-09-18" "2014-09-19" "2014-09-20"
#>  [261] "2014-09-21" "2014-09-22" "2014-09-23" "2014-09-24" "2014-09-25"
#>  [266] "2014-09-26" "2014-09-27" "2014-09-28" "2014-09-29" "2014-09-30"
#>  [271] "2014-10-01" "2014-10-02" "2014-10-03" "2014-10-04" "2014-10-05"
#>  [276] "2014-10-06" "2014-10-07" "2014-10-08" "2014-10-09" "2014-10-10"
#>  [281] "2014-10-11" "2014-10-12" "2014-10-13" "2014-10-14" "2014-10-15"
#>  [286] "2014-10-16" "2014-10-17" "2014-10-18" "2014-10-19" "2014-10-20"
#>  [291] "2014-10-21" "2014-10-22" "2014-10-23" "2014-10-24" "2014-10-25"
#>  [296] "2014-10-26" "2014-10-27" "2014-10-28" "2014-10-29" "2014-10-30"
#>  [301] "2014-11-01" "2014-11-02" "2014-11-03" "2014-11-04" "2014-11-05"
#>  [306] "2014-11-06" "2014-11-07" "2014-11-08" "2014-11-09" "2014-11-10"
#>  [311] "2014-11-11" "2014-11-12" "2014-11-13" "2014-11-14" "2014-11-15"
#>  [316] "2014-11-16" "2014-11-17" "2014-11-18" "2014-11-19" "2014-11-20"
#>  [321] "2014-11-21" "2014-11-22" "2014-11-23" "2014-11-24" "2014-11-25"
#>  [326] "2014-11-26" "2014-11-27" "2014-11-28" "2014-11-29" "2014-11-30"
#>  [331] "2014-12-01" "2014-12-02" "2014-12-03" "2014-12-04" "2014-12-05"
#>  [336] "2014-12-06" "2014-12-07" "2014-12-08" "2014-12-09" "2014-12-10"
#>  [341] "2014-12-11" "2014-12-12" "2014-12-13" "2014-12-14" "2014-12-15"
#>  [346] "2014-12-16" "2014-12-17" "2014-12-18" "2014-12-19" "2014-12-20"
#>  [351] "2014-12-21" "2014-12-22" "2014-12-23" "2014-12-24" "2014-12-25"
#>  [356] "2014-12-26" "2014-12-27" "2014-12-28" "2014-12-29" "2014-12-30"
#>  [361] "2015-01-01" "2015-01-02" "2015-01-03" "2015-01-04" "2015-01-05"
#>  [366] "2015-01-06" "2015-01-07" "2015-01-08" "2015-01-09" "2015-01-10"
#>  [371] "2015-01-11" "2015-01-12" "2015-01-13" "2015-01-14" "2015-01-15"
#>  [376] "2015-01-16" "2015-01-17" "2015-01-18" "2015-01-19" "2015-01-20"
#>  [381] "2015-01-21" "2015-01-22" "2015-01-23" "2015-01-24" "2015-01-25"
#>  [386] "2015-01-26" "2015-01-27" "2015-01-28" "2015-01-29" "2015-01-30"
#>  [391] "2015-02-01" "2015-02-02" "2015-02-03" "2015-02-04" "2015-02-05"
#>  [396] "2015-02-06" "2015-02-07" "2015-02-08" "2015-02-09" "2015-02-10"
#>  [401] "2015-02-11" "2015-02-12" "2015-02-13" "2015-02-14" "2015-02-15"
#>  [406] "2015-02-16" "2015-02-17" "2015-02-18" "2015-02-19" "2015-02-20"
#>  [411] "2015-02-21" "2015-02-22" "2015-02-23" "2015-02-24" "2015-02-25"
#>  [416] "2015-02-26" "2015-02-27" "2015-02-28" "2015-02-29" "2015-02-30"
#>  [421] "2015-03-01" "2015-03-02" "2015-03-03" "2015-03-04" "2015-03-05"
#>  [426] "2015-03-06" "2015-03-07" "2015-03-08" "2015-03-09" "2015-03-10"
#>  [431] "2015-03-11" "2015-03-12" "2015-03-13" "2015-03-14" "2015-03-15"
#>  [436] "2015-03-16" "2015-03-17" "2015-03-18" "2015-03-19" "2015-03-20"
#>  [441] "2015-03-21" "2015-03-22" "2015-03-23" "2015-03-24" "2015-03-25"
#>  [446] "2015-03-26" "2015-03-27" "2015-03-28" "2015-03-29" "2015-03-30"
#>  [451] "2015-04-01" "2015-04-02" "2015-04-03" "2015-04-04" "2015-04-05"
#>  [456] "2015-04-06" "2015-04-07" "2015-04-08" "2015-04-09" "2015-04-10"
#>  [461] "2015-04-11" "2015-04-12" "2015-04-13" "2015-04-14" "2015-04-15"
#>  [466] "2015-04-16" "2015-04-17" "2015-04-18" "2015-04-19" "2015-04-20"
#>  [471] "2015-04-21" "2015-04-22" "2015-04-23" "2015-04-24" "2015-04-25"
#>  [476] "2015-04-26" "2015-04-27" "2015-04-28" "2015-04-29" "2015-04-30"
#>  [481] "2015-05-01" "2015-05-02" "2015-05-03" "2015-05-04" "2015-05-05"
#>  [486] "2015-05-06" "2015-05-07" "2015-05-08" "2015-05-09" "2015-05-10"
#>  [491] "2015-05-11" "2015-05-12" "2015-05-13" "2015-05-14" "2015-05-15"
#>  [496] "2015-05-16" "2015-05-17" "2015-05-18" "2015-05-19" "2015-05-20"
#>  [501] "2015-05-21" "2015-05-22" "2015-05-23" "2015-05-24" "2015-05-25"
#>  [506] "2015-05-26" "2015-05-27" "2015-05-28" "2015-05-29" "2015-05-30"
#>  [511] "2015-06-01" "2015-06-02" "2015-06-03" "2015-06-04" "2015-06-05"
#>  [516] "2015-06-06" "2015-06-07" "2015-06-08" "2015-06-09" "2015-06-10"
#>  [521] "2015-06-11" "2015-06-12" "2015-06-13" "2015-06-14" "2015-06-15"
#>  [526] "2015-06-16" "2015-06-17" "2015-06-18" "2015-06-19" "2015-06-20"
#>  [531] "2015-06-21" "2015-06-22" "2015-06-23" "2015-06-24" "2015-06-25"
#>  [536] "2015-06-26" "2015-06-27" "2015-06-28" "2015-06-29" "2015-06-30"
#>  [541] "2015-07-01" "2015-07-02" "2015-07-03" "2015-07-04" "2015-07-05"
#>  [546] "2015-07-06" "2015-07-07" "2015-07-08" "2015-07-09" "2015-07-10"
#>  [551] "2015-07-11" "2015-07-12" "2015-07-13" "2015-07-14" "2015-07-15"
#>  [556] "2015-07-16" "2015-07-17" "2015-07-18" "2015-07-19" "2015-07-20"
#>  [561] "2015-07-21" "2015-07-22" "2015-07-23" "2015-07-24" "2015-07-25"
#>  [566] "2015-07-26" "2015-07-27" "2015-07-28" "2015-07-29" "2015-07-30"
#>  [571] "2015-08-01" "2015-08-02" "2015-08-03" "2015-08-04" "2015-08-05"
#>  [576] "2015-08-06" "2015-08-07" "2015-08-08" "2015-08-09" "2015-08-10"
#>  [581] "2015-08-11" "2015-08-12" "2015-08-13" "2015-08-14" "2015-08-15"
#>  [586] "2015-08-16" "2015-08-17" "2015-08-18" "2015-08-19" "2015-08-20"
#>  [591] "2015-08-21" "2015-08-22" "2015-08-23" "2015-08-24" "2015-08-25"
#>  [596] "2015-08-26" "2015-08-27" "2015-08-28" "2015-08-29" "2015-08-30"
#>  [601] "2015-09-01" "2015-09-02" "2015-09-03" "2015-09-04" "2015-09-05"
#>  [606] "2015-09-06" "2015-09-07" "2015-09-08" "2015-09-09" "2015-09-10"
#>  [611] "2015-09-11" "2015-09-12" "2015-09-13" "2015-09-14" "2015-09-15"
#>  [616] "2015-09-16" "2015-09-17" "2015-09-18" "2015-09-19" "2015-09-20"
#>  [621] "2015-09-21" "2015-09-22" "2015-09-23" "2015-09-24" "2015-09-25"
#>  [626] "2015-09-26" "2015-09-27" "2015-09-28" "2015-09-29" "2015-09-30"
#>  [631] "2015-10-01" "2015-10-02" "2015-10-03" "2015-10-04" "2015-10-05"
#>  [636] "2015-10-06" "2015-10-07" "2015-10-08" "2015-10-09" "2015-10-10"
#>  [641] "2015-10-11" "2015-10-12" "2015-10-13" "2015-10-14" "2015-10-15"
#>  [646] "2015-10-16" "2015-10-17" "2015-10-18" "2015-10-19" "2015-10-20"
#>  [651] "2015-10-21" "2015-10-22" "2015-10-23" "2015-10-24" "2015-10-25"
#>  [656] "2015-10-26" "2015-10-27" "2015-10-28" "2015-10-29" "2015-10-30"
#>  [661] "2015-11-01" "2015-11-02" "2015-11-03" "2015-11-04" "2015-11-05"
#>  [666] "2015-11-06" "2015-11-07" "2015-11-08" "2015-11-09" "2015-11-10"
#>  [671] "2015-11-11" "2015-11-12" "2015-11-13" "2015-11-14" "2015-11-15"
#>  [676] "2015-11-16" "2015-11-17" "2015-11-18" "2015-11-19" "2015-11-20"
#>  [681] "2015-11-21" "2015-11-22" "2015-11-23" "2015-11-24" "2015-11-25"
#>  [686] "2015-11-26" "2015-11-27" "2015-11-28" "2015-11-29" "2015-11-30"
#>  [691] "2015-12-01" "2015-12-02" "2015-12-03" "2015-12-04" "2015-12-05"
#>  [696] "2015-12-06" "2015-12-07" "2015-12-08" "2015-12-09" "2015-12-10"
#>  [701] "2015-12-11" "2015-12-12" "2015-12-13" "2015-12-14" "2015-12-15"
#>  [706] "2015-12-16" "2015-12-17" "2015-12-18" "2015-12-19" "2015-12-20"
#>  [711] "2015-12-21" "2015-12-22" "2015-12-23" "2015-12-24" "2015-12-25"
#>  [716] "2015-12-26" "2015-12-27" "2015-12-28" "2015-12-29" "2015-12-30"
#>  [721] "2016-01-01" "2016-01-02" "2016-01-03" "2016-01-04" "2016-01-05"
#>  [726] "2016-01-06" "2016-01-07" "2016-01-08" "2016-01-09" "2016-01-10"
#>  [731] "2016-01-11" "2016-01-12" "2016-01-13" "2016-01-14" "2016-01-15"
#>  [736] "2016-01-16" "2016-01-17" "2016-01-18" "2016-01-19" "2016-01-20"
#>  [741] "2016-01-21" "2016-01-22" "2016-01-23" "2016-01-24" "2016-01-25"
#>  [746] "2016-01-26" "2016-01-27" "2016-01-28" "2016-01-29" "2016-01-30"
#>  [751] "2016-02-01" "2016-02-02" "2016-02-03" "2016-02-04" "2016-02-05"
#>  [756] "2016-02-06" "2016-02-07" "2016-02-08" "2016-02-09" "2016-02-10"
#>  [761] "2016-02-11" "2016-02-12" "2016-02-13" "2016-02-14" "2016-02-15"
#>  [766] "2016-02-16" "2016-02-17" "2016-02-18" "2016-02-19" "2016-02-20"
#>  [771] "2016-02-21" "2016-02-22" "2016-02-23" "2016-02-24" "2016-02-25"
#>  [776] "2016-02-26" "2016-02-27" "2016-02-28" "2016-02-29" "2016-02-30"
#>  [781] "2016-03-01" "2016-03-02" "2016-03-03" "2016-03-04" "2016-03-05"
#>  [786] "2016-03-06" "2016-03-07" "2016-03-08" "2016-03-09" "2016-03-10"
#>  [791] "2016-03-11" "2016-03-12" "2016-03-13" "2016-03-14" "2016-03-15"
#>  [796] "2016-03-16" "2016-03-17" "2016-03-18" "2016-03-19" "2016-03-20"
#>  [801] "2016-03-21" "2016-03-22" "2016-03-23" "2016-03-24" "2016-03-25"
#>  [806] "2016-03-26" "2016-03-27" "2016-03-28" "2016-03-29" "2016-03-30"
#>  [811] "2016-04-01" "2016-04-02" "2016-04-03" "2016-04-04" "2016-04-05"
#>  [816] "2016-04-06" "2016-04-07" "2016-04-08" "2016-04-09" "2016-04-10"
#>  [821] "2016-04-11" "2016-04-12" "2016-04-13" "2016-04-14" "2016-04-15"
#>  [826] "2016-04-16" "2016-04-17" "2016-04-18" "2016-04-19" "2016-04-20"
#>  [831] "2016-04-21" "2016-04-22" "2016-04-23" "2016-04-24" "2016-04-25"
#>  [836] "2016-04-26" "2016-04-27" "2016-04-28" "2016-04-29" "2016-04-30"
#>  [841] "2016-05-01" "2016-05-02" "2016-05-03" "2016-05-04" "2016-05-05"
#>  [846] "2016-05-06" "2016-05-07" "2016-05-08" "2016-05-09" "2016-05-10"
#>  [851] "2016-05-11" "2016-05-12" "2016-05-13" "2016-05-14" "2016-05-15"
#>  [856] "2016-05-16" "2016-05-17" "2016-05-18" "2016-05-19" "2016-05-20"
#>  [861] "2016-05-21" "2016-05-22" "2016-05-23" "2016-05-24" "2016-05-25"
#>  [866] "2016-05-26" "2016-05-27" "2016-05-28" "2016-05-29" "2016-05-30"
#>  [871] "2016-06-01" "2016-06-02" "2016-06-03" "2016-06-04" "2016-06-05"
#>  [876] "2016-06-06" "2016-06-07" "2016-06-08" "2016-06-09" "2016-06-10"
#>  [881] "2016-06-11" "2016-06-12" "2016-06-13" "2016-06-14" "2016-06-15"
#>  [886] "2016-06-16" "2016-06-17" "2016-06-18" "2016-06-19" "2016-06-20"
#>  [891] "2016-06-21" "2016-06-22" "2016-06-23" "2016-06-24" "2016-06-25"
#>  [896] "2016-06-26" "2016-06-27" "2016-06-28" "2016-06-29" "2016-06-30"
#>  [901] "2016-07-01" "2016-07-02" "2016-07-03" "2016-07-04" "2016-07-05"
#>  [906] "2016-07-06" "2016-07-07" "2016-07-08" "2016-07-09" "2016-07-10"
#>  [911] "2016-07-11" "2016-07-12" "2016-07-13" "2016-07-14" "2016-07-15"
#>  [916] "2016-07-16" "2016-07-17" "2016-07-18" "2016-07-19" "2016-07-20"
#>  [921] "2016-07-21" "2016-07-22" "2016-07-23" "2016-07-24" "2016-07-25"
#>  [926] "2016-07-26" "2016-07-27" "2016-07-28" "2016-07-29" "2016-07-30"
#>  [931] "2016-08-01" "2016-08-02" "2016-08-03" "2016-08-04" "2016-08-05"
#>  [936] "2016-08-06" "2016-08-07" "2016-08-08" "2016-08-09" "2016-08-10"
#>  [941] "2016-08-11" "2016-08-12" "2016-08-13" "2016-08-14" "2016-08-15"
#>  [946] "2016-08-16" "2016-08-17" "2016-08-18" "2016-08-19" "2016-08-20"
#>  [951] "2016-08-21" "2016-08-22" "2016-08-23" "2016-08-24" "2016-08-25"
#>  [956] "2016-08-26" "2016-08-27" "2016-08-28" "2016-08-29" "2016-08-30"
#>  [961] "2016-09-01" "2016-09-02" "2016-09-03" "2016-09-04" "2016-09-05"
#>  [966] "2016-09-06" "2016-09-07" "2016-09-08" "2016-09-09" "2016-09-10"
#>  [971] "2016-09-11" "2016-09-12" "2016-09-13" "2016-09-14" "2016-09-15"
#>  [976] "2016-09-16" "2016-09-17" "2016-09-18" "2016-09-19" "2016-09-20"
#>  [981] "2016-09-21" "2016-09-22" "2016-09-23" "2016-09-24" "2016-09-25"
#>  [986] "2016-09-26" "2016-09-27" "2016-09-28" "2016-09-29" "2016-09-30"
#>  [991] "2016-10-01" "2016-10-02" "2016-10-03" "2016-10-04" "2016-10-05"
#>  [996] "2016-10-06" "2016-10-07" "2016-10-08" "2016-10-09" "2016-10-10"
#> [1001] "2016-10-11" "2016-10-12" "2016-10-13" "2016-10-14" "2016-10-15"
#> [1006] "2016-10-16" "2016-10-17" "2016-10-18" "2016-10-19" "2016-10-20"
#> [1011] "2016-10-21" "2016-10-22" "2016-10-23" "2016-10-24" "2016-10-25"
#> [1016] "2016-10-26" "2016-10-27" "2016-10-28" "2016-10-29" "2016-10-30"
#> [1021] "2016-11-01" "2016-11-02" "2016-11-03" "2016-11-04" "2016-11-05"
#> [1026] "2016-11-06" "2016-11-07" "2016-11-08" "2016-11-09" "2016-11-10"
#> [1031] "2016-11-11" "2016-11-12" "2016-11-13" "2016-11-14" "2016-11-15"
#> [1036] "2016-11-16" "2016-11-17" "2016-11-18" "2016-11-19" "2016-11-20"
#> [1041] "2016-11-21" "2016-11-22" "2016-11-23" "2016-11-24" "2016-11-25"
#> [1046] "2016-11-26" "2016-11-27" "2016-11-28" "2016-11-29" "2016-11-30"
#> [1051] "2016-12-01" "2016-12-02" "2016-12-03" "2016-12-04" "2016-12-05"
#> [1056] "2016-12-06" "2016-12-07" "2016-12-08" "2016-12-09" "2016-12-10"
#> [1061] "2016-12-11" "2016-12-12" "2016-12-13" "2016-12-14" "2016-12-15"
#> [1066] "2016-12-16" "2016-12-17" "2016-12-18" "2016-12-19" "2016-12-20"
#> [1071] "2016-12-21" "2016-12-22" "2016-12-23" "2016-12-24" "2016-12-25"
#> [1076] "2016-12-26" "2016-12-27" "2016-12-28" "2016-12-29" "2016-12-30"
#> [1081] "2017-01-01" "2017-01-02" "2017-01-03" "2017-01-04" "2017-01-05"
#> [1086] "2017-01-06" "2017-01-07" "2017-01-08" "2017-01-09" "2017-01-10"
#> [1091] "2017-01-11" "2017-01-12" "2017-01-13" "2017-01-14" "2017-01-15"
#> [1096] "2017-01-16" "2017-01-17" "2017-01-18" "2017-01-19" "2017-01-20"
#> [1101] "2017-01-21" "2017-01-22" "2017-01-23" "2017-01-24" "2017-01-25"
#> [1106] "2017-01-26" "2017-01-27" "2017-01-28" "2017-01-29" "2017-01-30"
#> [1111] "2017-02-01" "2017-02-02" "2017-02-03" "2017-02-04" "2017-02-05"
#> [1116] "2017-02-06" "2017-02-07" "2017-02-08" "2017-02-09" "2017-02-10"
#> [1121] "2017-02-11" "2017-02-12" "2017-02-13" "2017-02-14" "2017-02-15"
#> [1126] "2017-02-16" "2017-02-17" "2017-02-18" "2017-02-19" "2017-02-20"
#> [1131] "2017-02-21" "2017-02-22" "2017-02-23" "2017-02-24" "2017-02-25"
#> [1136] "2017-02-26" "2017-02-27" "2017-02-28" "2017-02-29" "2017-02-30"
#> [1141] "2017-03-01" "2017-03-02" "2017-03-03" "2017-03-04" "2017-03-05"
#> [1146] "2017-03-06" "2017-03-07" "2017-03-08" "2017-03-09" "2017-03-10"
#> [1151] "2017-03-11" "2017-03-12" "2017-03-13" "2017-03-14" "2017-03-15"
#> [1156] "2017-03-16" "2017-03-17" "2017-03-18" "2017-03-19" "2017-03-20"
#> [1161] "2017-03-21" "2017-03-22" "2017-03-23" "2017-03-24" "2017-03-25"
#> [1166] "2017-03-26" "2017-03-27" "2017-03-28" "2017-03-29" "2017-03-30"
#> [1171] "2017-04-01" "2017-04-02" "2017-04-03" "2017-04-04" "2017-04-05"
#> [1176] "2017-04-06" "2017-04-07" "2017-04-08" "2017-04-09" "2017-04-10"
#> [1181] "2017-04-11" "2017-04-12" "2017-04-13" "2017-04-14" "2017-04-15"
#> [1186] "2017-04-16" "2017-04-17" "2017-04-18" "2017-04-19" "2017-04-20"
#> [1191] "2017-04-21" "2017-04-22" "2017-04-23" "2017-04-24" "2017-04-25"
#> [1196] "2017-04-26" "2017-04-27" "2017-04-28" "2017-04-29" "2017-04-30"
#> [1201] "2017-05-01" "2017-05-02" "2017-05-03" "2017-05-04" "2017-05-05"
#> [1206] "2017-05-06" "2017-05-07" "2017-05-08" "2017-05-09" "2017-05-10"
#> [1211] "2017-05-11" "2017-05-12" "2017-05-13" "2017-05-14" "2017-05-15"
#> [1216] "2017-05-16" "2017-05-17" "2017-05-18" "2017-05-19" "2017-05-20"
#> [1221] "2017-05-21" "2017-05-22" "2017-05-23" "2017-05-24" "2017-05-25"
#> [1226] "2017-05-26" "2017-05-27" "2017-05-28" "2017-05-29" "2017-05-30"
#> [1231] "2017-06-01" "2017-06-02" "2017-06-03" "2017-06-04" "2017-06-05"
#> [1236] "2017-06-06" "2017-06-07" "2017-06-08" "2017-06-09" "2017-06-10"
#> [1241] "2017-06-11" "2017-06-12" "2017-06-13" "2017-06-14" "2017-06-15"
#> [1246] "2017-06-16" "2017-06-17" "2017-06-18" "2017-06-19" "2017-06-20"
#> [1251] "2017-06-21" "2017-06-22" "2017-06-23" "2017-06-24" "2017-06-25"
#> [1256] "2017-06-26" "2017-06-27" "2017-06-28" "2017-06-29" "2017-06-30"
#> [1261] "2017-07-01" "2017-07-02" "2017-07-03" "2017-07-04" "2017-07-05"
#> [1266] "2017-07-06" "2017-07-07" "2017-07-08" "2017-07-09" "2017-07-10"
#> [1271] "2017-07-11" "2017-07-12" "2017-07-13" "2017-07-14" "2017-07-15"
#> [1276] "2017-07-16" "2017-07-17" "2017-07-18" "2017-07-19" "2017-07-20"
#> [1281] "2017-07-21" "2017-07-22" "2017-07-23" "2017-07-24" "2017-07-25"
#> [1286] "2017-07-26" "2017-07-27" "2017-07-28" "2017-07-29" "2017-07-30"
#> [1291] "2017-08-01" "2017-08-02" "2017-08-03" "2017-08-04" "2017-08-05"
#> [1296] "2017-08-06" "2017-08-07" "2017-08-08" "2017-08-09" "2017-08-10"
#> [1301] "2017-08-11" "2017-08-12" "2017-08-13" "2017-08-14" "2017-08-15"
#> [1306] "2017-08-16" "2017-08-17" "2017-08-18" "2017-08-19" "2017-08-20"
#> [1311] "2017-08-21" "2017-08-22" "2017-08-23" "2017-08-24" "2017-08-25"
#> [1316] "2017-08-26" "2017-08-27" "2017-08-28" "2017-08-29" "2017-08-30"
#> [1321] "2017-09-01" "2017-09-02" "2017-09-03" "2017-09-04" "2017-09-05"
#> [1326] "2017-09-06" "2017-09-07" "2017-09-08" "2017-09-09" "2017-09-10"
#> [1331] "2017-09-11" "2017-09-12" "2017-09-13" "2017-09-14" "2017-09-15"
#> [1336] "2017-09-16" "2017-09-17" "2017-09-18" "2017-09-19" "2017-09-20"
#> [1341] "2017-09-21" "2017-09-22" "2017-09-23" "2017-09-24" "2017-09-25"
#> [1346] "2017-09-26" "2017-09-27" "2017-09-28" "2017-09-29" "2017-09-30"
#> [1351] "2017-10-01" "2017-10-02" "2017-10-03" "2017-10-04" "2017-10-05"
#> [1356] "2017-10-06" "2017-10-07" "2017-10-08" "2017-10-09" "2017-10-10"
#> [1361] "2017-10-11" "2017-10-12" "2017-10-13" "2017-10-14" "2017-10-15"
#> [1366] "2017-10-16" "2017-10-17" "2017-10-18" "2017-10-19" "2017-10-20"
#> [1371] "2017-10-21" "2017-10-22" "2017-10-23" "2017-10-24" "2017-10-25"
#> [1376] "2017-10-26" "2017-10-27" "2017-10-28" "2017-10-29" "2017-10-30"
#> [1381] "2017-11-01" "2017-11-02" "2017-11-03" "2017-11-04" "2017-11-05"
#> [1386] "2017-11-06" "2017-11-07" "2017-11-08" "2017-11-09" "2017-11-10"
#> [1391] "2017-11-11" "2017-11-12" "2017-11-13" "2017-11-14" "2017-11-15"
#> [1396] "2017-11-16" "2017-11-17" "2017-11-18" "2017-11-19" "2017-11-20"
#> [1401] "2017-11-21" "2017-11-22" "2017-11-23" "2017-11-24" "2017-11-25"
#> [1406] "2017-11-26" "2017-11-27" "2017-11-28" "2017-11-29" "2017-11-30"
#> [1411] "2017-12-01" "2017-12-02" "2017-12-03" "2017-12-04" "2017-12-05"
#> [1416] "2017-12-06" "2017-12-07" "2017-12-08" "2017-12-09" "2017-12-10"
#> [1421] "2017-12-11" "2017-12-12" "2017-12-13" "2017-12-14" "2017-12-15"
#> [1426] "2017-12-16" "2017-12-17" "2017-12-18" "2017-12-19" "2017-12-20"
#> [1431] "2017-12-21" "2017-12-22" "2017-12-23" "2017-12-24" "2017-12-25"
#> [1436] "2017-12-26" "2017-12-27" "2017-12-28" "2017-12-29" "2017-12-30"
#> [1441] "2018-01-01" "2018-01-02" "2018-01-03" "2018-01-04" "2018-01-05"
#> [1446] "2018-01-06" "2018-01-07" "2018-01-08" "2018-01-09" "2018-01-10"
#> [1451] "2018-01-11" "2018-01-12" "2018-01-13" "2018-01-14" "2018-01-15"
#> [1456] "2018-01-16" "2018-01-17" "2018-01-18" "2018-01-19" "2018-01-20"
#> [1461] "2018-01-21" "2018-01-22" "2018-01-23" "2018-01-24" "2018-01-25"
#> [1466] "2018-01-26" "2018-01-27" "2018-01-28" "2018-01-29" "2018-01-30"
#> [1471] "2018-02-01" "2018-02-02" "2018-02-03" "2018-02-04" "2018-02-05"
#> [1476] "2018-02-06" "2018-02-07" "2018-02-08" "2018-02-09" "2018-02-10"
#> [1481] "2018-02-11" "2018-02-12" "2018-02-13" "2018-02-14" "2018-02-15"
#> [1486] "2018-02-16" "2018-02-17" "2018-02-18" "2018-02-19" "2018-02-20"
#> [1491] "2018-02-21" "2018-02-22" "2018-02-23" "2018-02-24" "2018-02-25"
#> [1496] "2018-02-26" "2018-02-27" "2018-02-28" "2018-02-29" "2018-02-30"
#> [1501] "2018-03-01" "2018-03-02" "2018-03-03" "2018-03-04" "2018-03-05"
#> [1506] "2018-03-06" "2018-03-07" "2018-03-08" "2018-03-09" "2018-03-10"
#> [1511] "2018-03-11" "2018-03-12" "2018-03-13" "2018-03-14" "2018-03-15"
#> [1516] "2018-03-16" "2018-03-17" "2018-03-18" "2018-03-19" "2018-03-20"
#> [1521] "2018-03-21" "2018-03-22" "2018-03-23" "2018-03-24" "2018-03-25"
#> [1526] "2018-03-26" "2018-03-27" "2018-03-28" "2018-03-29" "2018-03-30"
#> [1531] "2018-04-01" "2018-04-02" "2018-04-03" "2018-04-04" "2018-04-05"
#> [1536] "2018-04-06" "2018-04-07" "2018-04-08" "2018-04-09" "2018-04-10"
#> [1541] "2018-04-11" "2018-04-12" "2018-04-13" "2018-04-14" "2018-04-15"
#> [1546] "2018-04-16" "2018-04-17" "2018-04-18" "2018-04-19" "2018-04-20"
#> [1551] "2018-04-21" "2018-04-22" "2018-04-23" "2018-04-24" "2018-04-25"
#> [1556] "2018-04-26" "2018-04-27" "2018-04-28" "2018-04-29" "2018-04-30"
#> [1561] "2018-05-01" "2018-05-02" "2018-05-03" "2018-05-04" "2018-05-05"
#> [1566] "2018-05-06" "2018-05-07" "2018-05-08" "2018-05-09" "2018-05-10"
#> [1571] "2018-05-11" "2018-05-12" "2018-05-13" "2018-05-14" "2018-05-15"
#> [1576] "2018-05-16" "2018-05-17" "2018-05-18" "2018-05-19" "2018-05-20"
#> [1581] "2018-05-21" "2018-05-22" "2018-05-23" "2018-05-24" "2018-05-25"
#> [1586] "2018-05-26" "2018-05-27" "2018-05-28" "2018-05-29" "2018-05-30"
#> [1591] "2018-06-01" "2018-06-02" "2018-06-03" "2018-06-04" "2018-06-05"
#> [1596] "2018-06-06" "2018-06-07" "2018-06-08" "2018-06-09" "2018-06-10"
#> [1601] "2018-06-11" "2018-06-12" "2018-06-13" "2018-06-14" "2018-06-15"
#> [1606] "2018-06-16" "2018-06-17" "2018-06-18" "2018-06-19" "2018-06-20"
#> [1611] "2018-06-21" "2018-06-22" "2018-06-23" "2018-06-24" "2018-06-25"
#> [1616] "2018-06-26" "2018-06-27" "2018-06-28" "2018-06-29" "2018-06-30"
#> [1621] "2018-07-01" "2018-07-02" "2018-07-03" "2018-07-04" "2018-07-05"
#> [1626] "2018-07-06" "2018-07-07" "2018-07-08" "2018-07-09" "2018-07-10"
#> [1631] "2018-07-11" "2018-07-12" "2018-07-13" "2018-07-14" "2018-07-15"
#> [1636] "2018-07-16" "2018-07-17" "2018-07-18" "2018-07-19" "2018-07-20"
#> [1641] "2018-07-21" "2018-07-22" "2018-07-23" "2018-07-24" "2018-07-25"
#> [1646] "2018-07-26" "2018-07-27" "2018-07-28" "2018-07-29" "2018-07-30"
#> [1651] "2018-08-01" "2018-08-02" "2018-08-03" "2018-08-04" "2018-08-05"
#> [1656] "2018-08-06" "2018-08-07" "2018-08-08" "2018-08-09" "2018-08-10"
#> [1661] "2018-08-11" "2018-08-12" "2018-08-13" "2018-08-14" "2018-08-15"
#> [1666] "2018-08-16" "2018-08-17" "2018-08-18" "2018-08-19" "2018-08-20"
#> [1671] "2018-08-21" "2018-08-22" "2018-08-23" "2018-08-24" "2018-08-25"
#> [1676] "2018-08-26" "2018-08-27" "2018-08-28" "2018-08-29" "2018-08-30"
#> [1681] "2018-09-01" "2018-09-02" "2018-09-03" "2018-09-04" "2018-09-05"
#> [1686] "2018-09-06" "2018-09-07" "2018-09-08" "2018-09-09" "2018-09-10"
#> [1691] "2018-09-11" "2018-09-12" "2018-09-13" "2018-09-14" "2018-09-15"
#> [1696] "2018-09-16" "2018-09-17" "2018-09-18" "2018-09-19" "2018-09-20"
#> [1701] "2018-09-21" "2018-09-22" "2018-09-23" "2018-09-24" "2018-09-25"
#> [1706] "2018-09-26" "2018-09-27" "2018-09-28" "2018-09-29" "2018-09-30"
#> [1711] "2018-10-01" "2018-10-02" "2018-10-03" "2018-10-04" "2018-10-05"
#> [1716] "2018-10-06" "2018-10-07" "2018-10-08" "2018-10-09" "2018-10-10"
#> [1721] "2018-10-11" "2018-10-12" "2018-10-13" "2018-10-14" "2018-10-15"
#> [1726] "2018-10-16" "2018-10-17" "2018-10-18" "2018-10-19" "2018-10-20"
#> [1731] "2018-10-21" "2018-10-22" "2018-10-23" "2018-10-24" "2018-10-25"
#> [1736] "2018-10-26" "2018-10-27" "2018-10-28" "2018-10-29" "2018-10-30"
#> [1741] "2018-11-01" "2018-11-02" "2018-11-03" "2018-11-04" "2018-11-05"
#> [1746] "2018-11-06" "2018-11-07" "2018-11-08" "2018-11-09" "2018-11-10"
#> [1751] "2018-11-11" "2018-11-12" "2018-11-13" "2018-11-14" "2018-11-15"
#> [1756] "2018-11-16" "2018-11-17" "2018-11-18" "2018-11-19" "2018-11-20"
#> [1761] "2018-11-21" "2018-11-22" "2018-11-23" "2018-11-24" "2018-11-25"
#> [1766] "2018-11-26" "2018-11-27" "2018-11-28" "2018-11-29" "2018-11-30"
#> [1771] "2018-12-01" "2018-12-02" "2018-12-03" "2018-12-04" "2018-12-05"
#> [1776] "2018-12-06" "2018-12-07" "2018-12-08" "2018-12-09" "2018-12-10"
#> [1781] "2018-12-11" "2018-12-12" "2018-12-13" "2018-12-14" "2018-12-15"
#> [1786] "2018-12-16" "2018-12-17" "2018-12-18" "2018-12-19" "2018-12-20"
#> [1791] "2018-12-21" "2018-12-22" "2018-12-23" "2018-12-24" "2018-12-25"
#> [1796] "2018-12-26" "2018-12-27" "2018-12-28" "2018-12-29" "2018-12-30"
#> [1801] "2019-01-01" "2019-01-02" "2019-01-03" "2019-01-04" "2019-01-05"
#> [1806] "2019-01-06" "2019-01-07" "2019-01-08" "2019-01-09" "2019-01-10"
#> [1811] "2019-01-11" "2019-01-12" "2019-01-13" "2019-01-14" "2019-01-15"
#> [1816] "2019-01-16" "2019-01-17" "2019-01-18" "2019-01-19" "2019-01-20"
#> [1821] "2019-01-21" "2019-01-22" "2019-01-23" "2019-01-24" "2019-01-25"
#> [1826] "2019-01-26" "2019-01-27" "2019-01-28" "2019-01-29" "2019-01-30"
#> [1831] "2019-02-01" "2019-02-02" "2019-02-03" "2019-02-04" "2019-02-05"
#> [1836] "2019-02-06" "2019-02-07" "2019-02-08" "2019-02-09" "2019-02-10"
#> [1841] "2019-02-11" "2019-02-12" "2019-02-13" "2019-02-14" "2019-02-15"
#> [1846] "2019-02-16" "2019-02-17" "2019-02-18" "2019-02-19" "2019-02-20"
#> [1851] "2019-02-21" "2019-02-22" "2019-02-23" "2019-02-24" "2019-02-25"
#> [1856] "2019-02-26" "2019-02-27" "2019-02-28" "2019-02-29" "2019-02-30"
#> [1861] "2019-03-01" "2019-03-02" "2019-03-03" "2019-03-04" "2019-03-05"
#> [1866] "2019-03-06" "2019-03-07" "2019-03-08" "2019-03-09" "2019-03-10"
#> [1871] "2019-03-11" "2019-03-12" "2019-03-13" "2019-03-14" "2019-03-15"
#> [1876] "2019-03-16" "2019-03-17" "2019-03-18" "2019-03-19" "2019-03-20"
#> [1881] "2019-03-21" "2019-03-22" "2019-03-23" "2019-03-24" "2019-03-25"
#> [1886] "2019-03-26" "2019-03-27" "2019-03-28" "2019-03-29" "2019-03-30"
#> [1891] "2019-04-01" "2019-04-02" "2019-04-03" "2019-04-04" "2019-04-05"
#> [1896] "2019-04-06" "2019-04-07" "2019-04-08" "2019-04-09" "2019-04-10"
#> [1901] "2019-04-11" "2019-04-12" "2019-04-13" "2019-04-14" "2019-04-15"
#> [1906] "2019-04-16" "2019-04-17" "2019-04-18" "2019-04-19" "2019-04-20"
#> [1911] "2019-04-21" "2019-04-22" "2019-04-23" "2019-04-24" "2019-04-25"
#> [1916] "2019-04-26" "2019-04-27" "2019-04-28" "2019-04-29" "2019-04-30"
#> [1921] "2019-05-01" "2019-05-02" "2019-05-03" "2019-05-04" "2019-05-05"
#> [1926] "2019-05-06" "2019-05-07" "2019-05-08" "2019-05-09" "2019-05-10"
#> [1931] "2019-05-11" "2019-05-12" "2019-05-13" "2019-05-14" "2019-05-15"
#> [1936] "2019-05-16" "2019-05-17" "2019-05-18" "2019-05-19" "2019-05-20"
#> [1941] "2019-05-21" "2019-05-22" "2019-05-23" "2019-05-24" "2019-05-25"
#> [1946] "2019-05-26" "2019-05-27" "2019-05-28" "2019-05-29" "2019-05-30"
#> [1951] "2019-06-01" "2019-06-02" "2019-06-03" "2019-06-04" "2019-06-05"
#> [1956] "2019-06-06" "2019-06-07" "2019-06-08" "2019-06-09" "2019-06-10"
#> [1961] "2019-06-11" "2019-06-12" "2019-06-13" "2019-06-14" "2019-06-15"
#> [1966] "2019-06-16" "2019-06-17" "2019-06-18" "2019-06-19" "2019-06-20"
#> [1971] "2019-06-21" "2019-06-22" "2019-06-23" "2019-06-24" "2019-06-25"
#> [1976] "2019-06-26" "2019-06-27" "2019-06-28" "2019-06-29" "2019-06-30"
#> [1981] "2019-07-01" "2019-07-02" "2019-07-03" "2019-07-04" "2019-07-05"
#> [1986] "2019-07-06" "2019-07-07" "2019-07-08" "2019-07-09" "2019-07-10"
#> [1991] "2019-07-11" "2019-07-12" "2019-07-13" "2019-07-14" "2019-07-15"
#> [1996] "2019-07-16" "2019-07-17" "2019-07-18" "2019-07-19" "2019-07-20"
#> [2001] "2019-07-21" "2019-07-22" "2019-07-23" "2019-07-24" "2019-07-25"
#> [2006] "2019-07-26" "2019-07-27" "2019-07-28" "2019-07-29" "2019-07-30"
#> [2011] "2019-08-01" "2019-08-02" "2019-08-03" "2019-08-04" "2019-08-05"
#> [2016] "2019-08-06" "2019-08-07" "2019-08-08" "2019-08-09" "2019-08-10"
#> [2021] "2019-08-11" "2019-08-12" "2019-08-13" "2019-08-14" "2019-08-15"
#> [2026] "2019-08-16" "2019-08-17" "2019-08-18" "2019-08-19" "2019-08-20"
#> [2031] "2019-08-21" "2019-08-22" "2019-08-23" "2019-08-24" "2019-08-25"
#> [2036] "2019-08-26" "2019-08-27" "2019-08-28" "2019-08-29" "2019-08-30"
#> [2041] "2019-09-01" "2019-09-02" "2019-09-03" "2019-09-04" "2019-09-05"
#> [2046] "2019-09-06" "2019-09-07" "2019-09-08" "2019-09-09" "2019-09-10"
#> [2051] "2019-09-11" "2019-09-12" "2019-09-13" "2019-09-14" "2019-09-15"
#> [2056] "2019-09-16" "2019-09-17" "2019-09-18" "2019-09-19" "2019-09-20"
#> [2061] "2019-09-21" "2019-09-22" "2019-09-23" "2019-09-24" "2019-09-25"
#> [2066] "2019-09-26" "2019-09-27" "2019-09-28" "2019-09-29" "2019-09-30"
#> [2071] "2019-10-01" "2019-10-02" "2019-10-03" "2019-10-04" "2019-10-05"
#> [2076] "2019-10-06" "2019-10-07" "2019-10-08" "2019-10-09" "2019-10-10"
#> [2081] "2019-10-11" "2019-10-12" "2019-10-13" "2019-10-14" "2019-10-15"
#> [2086] "2019-10-16" "2019-10-17" "2019-10-18" "2019-10-19" "2019-10-20"
#> [2091] "2019-10-21" "2019-10-22" "2019-10-23" "2019-10-24" "2019-10-25"
#> [2096] "2019-10-26" "2019-10-27" "2019-10-28" "2019-10-29" "2019-10-30"
#> [2101] "2019-11-01" "2019-11-02" "2019-11-03" "2019-11-04" "2019-11-05"
#> [2106] "2019-11-06" "2019-11-07" "2019-11-08" "2019-11-09" "2019-11-10"
#> [2111] "2019-11-11" "2019-11-12" "2019-11-13" "2019-11-14" "2019-11-15"
#> [2116] "2019-11-16" "2019-11-17" "2019-11-18" "2019-11-19" "2019-11-20"
#> [2121] "2019-11-21" "2019-11-22" "2019-11-23" "2019-11-24" "2019-11-25"
#> [2126] "2019-11-26" "2019-11-27" "2019-11-28" "2019-11-29" "2019-11-30"
#> [2131] "2019-12-01" "2019-12-02" "2019-12-03" "2019-12-04" "2019-12-05"
#> [2136] "2019-12-06" "2019-12-07" "2019-12-08" "2019-12-09" "2019-12-10"
#> [2141] "2019-12-11" "2019-12-12" "2019-12-13" "2019-12-14" "2019-12-15"
#> [2146] "2019-12-16" "2019-12-17" "2019-12-18" "2019-12-19" "2019-12-20"
#> [2151] "2019-12-21" "2019-12-22" "2019-12-23" "2019-12-24" "2019-12-25"
#> [2156] "2019-12-26" "2019-12-27" "2019-12-28" "2019-12-29" "2019-12-30"
#> [2161] "2020-01-01" "2020-01-02" "2020-01-03" "2020-01-04" "2020-01-05"
#> [2166] "2020-01-06" "2020-01-07" "2020-01-08" "2020-01-09" "2020-01-10"
#> [2171] "2020-01-11" "2020-01-12" "2020-01-13" "2020-01-14" "2020-01-15"
#> [2176] "2020-01-16" "2020-01-17" "2020-01-18" "2020-01-19" "2020-01-20"
#> [2181] "2020-01-21" "2020-01-22" "2020-01-23" "2020-01-24" "2020-01-25"
#> [2186] "2020-01-26" "2020-01-27" "2020-01-28" "2020-01-29" "2020-01-30"
#> [2191] "2020-02-01" "2020-02-02" "2020-02-03" "2020-02-04" "2020-02-05"
#> [2196] "2020-02-06" "2020-02-07" "2020-02-08" "2020-02-09" "2020-02-10"
#> [2201] "2020-02-11" "2020-02-12" "2020-02-13" "2020-02-14" "2020-02-15"
#> [2206] "2020-02-16" "2020-02-17" "2020-02-18" "2020-02-19" "2020-02-20"
#> [2211] "2020-02-21" "2020-02-22" "2020-02-23" "2020-02-24" "2020-02-25"
#> [2216] "2020-02-26" "2020-02-27" "2020-02-28" "2020-02-29" "2020-02-30"
#> [2221] "2020-03-01" "2020-03-02" "2020-03-03" "2020-03-04" "2020-03-05"
#> [2226] "2020-03-06" "2020-03-07" "2020-03-08" "2020-03-09" "2020-03-10"
#> [2231] "2020-03-11" "2020-03-12" "2020-03-13" "2020-03-14" "2020-03-15"
#> [2236] "2020-03-16" "2020-03-17" "2020-03-18" "2020-03-19" "2020-03-20"
#> [2241] "2020-03-21" "2020-03-22" "2020-03-23" "2020-03-24" "2020-03-25"
#> [2246] "2020-03-26" "2020-03-27" "2020-03-28" "2020-03-29" "2020-03-30"
#> [2251] "2020-04-01" "2020-04-02" "2020-04-03" "2020-04-04" "2020-04-05"
#> [2256] "2020-04-06" "2020-04-07" "2020-04-08" "2020-04-09" "2020-04-10"
#> [2261] "2020-04-11" "2020-04-12" "2020-04-13" "2020-04-14" "2020-04-15"
#> [2266] "2020-04-16" "2020-04-17" "2020-04-18" "2020-04-19" "2020-04-20"
#> [2271] "2020-04-21" "2020-04-22" "2020-04-23" "2020-04-24" "2020-04-25"
#> [2276] "2020-04-26" "2020-04-27" "2020-04-28" "2020-04-29" "2020-04-30"
#> [2281] "2020-05-01" "2020-05-02" "2020-05-03" "2020-05-04" "2020-05-05"
#> [2286] "2020-05-06" "2020-05-07" "2020-05-08" "2020-05-09" "2020-05-10"
#> [2291] "2020-05-11" "2020-05-12" "2020-05-13" "2020-05-14" "2020-05-15"
#> [2296] "2020-05-16" "2020-05-17" "2020-05-18" "2020-05-19" "2020-05-20"
#> [2301] "2020-05-21" "2020-05-22" "2020-05-23" "2020-05-24" "2020-05-25"
#> [2306] "2020-05-26" "2020-05-27" "2020-05-28" "2020-05-29" "2020-05-30"
#> [2311] "2020-06-01" "2020-06-02" "2020-06-03" "2020-06-04" "2020-06-05"
#> [2316] "2020-06-06" "2020-06-07" "2020-06-08" "2020-06-09" "2020-06-10"
#> [2321] "2020-06-11" "2020-06-12" "2020-06-13" "2020-06-14" "2020-06-15"
#> [2326] "2020-06-16" "2020-06-17" "2020-06-18" "2020-06-19" "2020-06-20"
#> [2331] "2020-06-21" "2020-06-22" "2020-06-23" "2020-06-24" "2020-06-25"
#> [2336] "2020-06-26" "2020-06-27" "2020-06-28" "2020-06-29" "2020-06-30"
#> [2341] "2020-07-01" "2020-07-02" "2020-07-03" "2020-07-04" "2020-07-05"
#> [2346] "2020-07-06" "2020-07-07" "2020-07-08" "2020-07-09" "2020-07-10"
#> [2351] "2020-07-11" "2020-07-12" "2020-07-13" "2020-07-14" "2020-07-15"
#> [2356] "2020-07-16" "2020-07-17" "2020-07-18" "2020-07-19" "2020-07-20"
#> [2361] "2020-07-21" "2020-07-22" "2020-07-23" "2020-07-24" "2020-07-25"
#> [2366] "2020-07-26" "2020-07-27" "2020-07-28" "2020-07-29" "2020-07-30"
#> [2371] "2020-08-01" "2020-08-02" "2020-08-03" "2020-08-04" "2020-08-05"
#> [2376] "2020-08-06" "2020-08-07" "2020-08-08" "2020-08-09" "2020-08-10"
#> [2381] "2020-08-11" "2020-08-12" "2020-08-13" "2020-08-14" "2020-08-15"
#> [2386] "2020-08-16" "2020-08-17" "2020-08-18" "2020-08-19" "2020-08-20"
#> [2391] "2020-08-21" "2020-08-22" "2020-08-23" "2020-08-24" "2020-08-25"
#> [2396] "2020-08-26" "2020-08-27" "2020-08-28" "2020-08-29" "2020-08-30"
#> [2401] "2020-09-01" "2020-09-02" "2020-09-03" "2020-09-04" "2020-09-05"
#> [2406] "2020-09-06" "2020-09-07" "2020-09-08" "2020-09-09" "2020-09-10"
#> [2411] "2020-09-11" "2020-09-12" "2020-09-13" "2020-09-14" "2020-09-15"
#> [2416] "2020-09-16" "2020-09-17" "2020-09-18" "2020-09-19" "2020-09-20"
#> [2421] "2020-09-21" "2020-09-22" "2020-09-23" "2020-09-24" "2020-09-25"
#> [2426] "2020-09-26" "2020-09-27" "2020-09-28" "2020-09-29" "2020-09-30"
#> [2431] "2020-10-01" "2020-10-02" "2020-10-03" "2020-10-04" "2020-10-05"
#> [2436] "2020-10-06" "2020-10-07" "2020-10-08" "2020-10-09" "2020-10-10"
#> [2441] "2020-10-11" "2020-10-12" "2020-10-13" "2020-10-14" "2020-10-15"
#> [2446] "2020-10-16" "2020-10-17" "2020-10-18" "2020-10-19" "2020-10-20"
#> [2451] "2020-10-21" "2020-10-22" "2020-10-23" "2020-10-24" "2020-10-25"
#> [2456] "2020-10-26" "2020-10-27" "2020-10-28" "2020-10-29" "2020-10-30"
#> [2461] "2020-11-01" "2020-11-02" "2020-11-03" "2020-11-04" "2020-11-05"
#> [2466] "2020-11-06" "2020-11-07" "2020-11-08" "2020-11-09" "2020-11-10"
#> [2471] "2020-11-11" "2020-11-12" "2020-11-13" "2020-11-14" "2020-11-15"
#> [2476] "2020-11-16" "2020-11-17" "2020-11-18" "2020-11-19" "2020-11-20"
#> [2481] "2020-11-21" "2020-11-22" "2020-11-23" "2020-11-24" "2020-11-25"
#> [2486] "2020-11-26" "2020-11-27" "2020-11-28" "2020-11-29" "2020-11-30"
#> [2491] "2020-12-01" "2020-12-02" "2020-12-03" "2020-12-04" "2020-12-05"
#> [2496] "2020-12-06" "2020-12-07" "2020-12-08" "2020-12-09" "2020-12-10"
#> [2501] "2020-12-11" "2020-12-12" "2020-12-13" "2020-12-14" "2020-12-15"
#> [2506] "2020-12-16" "2020-12-17" "2020-12-18" "2020-12-19" "2020-12-20"
#> [2511] "2020-12-21" "2020-12-22" "2020-12-23" "2020-12-24" "2020-12-25"
#> [2516] "2020-12-26" "2020-12-27" "2020-12-28" "2020-12-29" "2020-12-30"
#> [2521] "2021-01-01" "2021-01-02" "2021-01-03" "2021-01-04" "2021-01-05"
#> [2526] "2021-01-06" "2021-01-07" "2021-01-08" "2021-01-09" "2021-01-10"
#> [2531] "2021-01-11" "2021-01-12" "2021-01-13" "2021-01-14" "2021-01-15"
#> [2536] "2021-01-16" "2021-01-17" "2021-01-18" "2021-01-19" "2021-01-20"
#> [2541] "2021-01-21" "2021-01-22" "2021-01-23" "2021-01-24" "2021-01-25"
#> [2546] "2021-01-26" "2021-01-27" "2021-01-28" "2021-01-29" "2021-01-30"
#> [2551] "2021-02-01" "2021-02-02" "2021-02-03" "2021-02-04" "2021-02-05"
#> [2556] "2021-02-06" "2021-02-07" "2021-02-08" "2021-02-09" "2021-02-10"
#> [2561] "2021-02-11" "2021-02-12" "2021-02-13" "2021-02-14" "2021-02-15"
#> [2566] "2021-02-16" "2021-02-17" "2021-02-18" "2021-02-19" "2021-02-20"
#> [2571] "2021-02-21" "2021-02-22" "2021-02-23" "2021-02-24" "2021-02-25"
#> [2576] "2021-02-26" "2021-02-27" "2021-02-28" "2021-02-29" "2021-02-30"
#> [2581] "2021-03-01" "2021-03-02" "2021-03-03" "2021-03-04" "2021-03-05"
#> [2586] "2021-03-06" "2021-03-07" "2021-03-08" "2021-03-09" "2021-03-10"
#> [2591] "2021-03-11" "2021-03-12" "2021-03-13" "2021-03-14" "2021-03-15"
#> [2596] "2021-03-16" "2021-03-17" "2021-03-18" "2021-03-19" "2021-03-20"
#> [2601] "2021-03-21" "2021-03-22" "2021-03-23" "2021-03-24" "2021-03-25"
#> [2606] "2021-03-26" "2021-03-27" "2021-03-28" "2021-03-29" "2021-03-30"
#> [2611] "2021-04-01" "2021-04-02" "2021-04-03" "2021-04-04" "2021-04-05"
#> [2616] "2021-04-06" "2021-04-07" "2021-04-08" "2021-04-09" "2021-04-10"
#> [2621] "2021-04-11" "2021-04-12" "2021-04-13" "2021-04-14" "2021-04-15"
#> [2626] "2021-04-16" "2021-04-17" "2021-04-18" "2021-04-19" "2021-04-20"
#> [2631] "2021-04-21" "2021-04-22" "2021-04-23" "2021-04-24" "2021-04-25"
#> [2636] "2021-04-26" "2021-04-27" "2021-04-28" "2021-04-29" "2021-04-30"
#> [2641] "2021-05-01" "2021-05-02" "2021-05-03" "2021-05-04" "2021-05-05"
#> [2646] "2021-05-06" "2021-05-07" "2021-05-08" "2021-05-09" "2021-05-10"
#> [2651] "2021-05-11" "2021-05-12" "2021-05-13" "2021-05-14" "2021-05-15"
#> [2656] "2021-05-16" "2021-05-17" "2021-05-18" "2021-05-19" "2021-05-20"
#> [2661] "2021-05-21" "2021-05-22" "2021-05-23" "2021-05-24" "2021-05-25"
#> [2666] "2021-05-26" "2021-05-27" "2021-05-28" "2021-05-29" "2021-05-30"
#> [2671] "2021-06-01" "2021-06-02" "2021-06-03" "2021-06-04" "2021-06-05"
#> [2676] "2021-06-06" "2021-06-07" "2021-06-08" "2021-06-09" "2021-06-10"
#> [2681] "2021-06-11" "2021-06-12" "2021-06-13" "2021-06-14" "2021-06-15"
#> [2686] "2021-06-16" "2021-06-17" "2021-06-18" "2021-06-19" "2021-06-20"
#> [2691] "2021-06-21" "2021-06-22" "2021-06-23" "2021-06-24" "2021-06-25"
#> [2696] "2021-06-26" "2021-06-27" "2021-06-28" "2021-06-29" "2021-06-30"
#> [2701] "2021-07-01" "2021-07-02" "2021-07-03" "2021-07-04" "2021-07-05"
#> [2706] "2021-07-06" "2021-07-07" "2021-07-08" "2021-07-09" "2021-07-10"
#> [2711] "2021-07-11" "2021-07-12" "2021-07-13" "2021-07-14" "2021-07-15"
#> [2716] "2021-07-16" "2021-07-17" "2021-07-18" "2021-07-19" "2021-07-20"
#> [2721] "2021-07-21" "2021-07-22" "2021-07-23" "2021-07-24" "2021-07-25"
#> [2726] "2021-07-26" "2021-07-27" "2021-07-28" "2021-07-29" "2021-07-30"
#> [2731] "2021-08-01" "2021-08-02" "2021-08-03" "2021-08-04" "2021-08-05"
#> [2736] "2021-08-06" "2021-08-07" "2021-08-08" "2021-08-09" "2021-08-10"
#> [2741] "2021-08-11" "2021-08-12" "2021-08-13" "2021-08-14" "2021-08-15"
#> [2746] "2021-08-16" "2021-08-17" "2021-08-18" "2021-08-19" "2021-08-20"
#> [2751] "2021-08-21" "2021-08-22" "2021-08-23" "2021-08-24" "2021-08-25"
#> [2756] "2021-08-26" "2021-08-27" "2021-08-28" "2021-08-29" "2021-08-30"
#> [2761] "2021-09-01" "2021-09-02" "2021-09-03" "2021-09-04" "2021-09-05"
#> [2766] "2021-09-06" "2021-09-07" "2021-09-08" "2021-09-09" "2021-09-10"
#> [2771] "2021-09-11" "2021-09-12" "2021-09-13" "2021-09-14" "2021-09-15"
#> [2776] "2021-09-16" "2021-09-17" "2021-09-18" "2021-09-19" "2021-09-20"
#> [2781] "2021-09-21" "2021-09-22" "2021-09-23" "2021-09-24" "2021-09-25"
#> [2786] "2021-09-26" "2021-09-27" "2021-09-28" "2021-09-29" "2021-09-30"
#> [2791] "2021-10-01" "2021-10-02" "2021-10-03" "2021-10-04" "2021-10-05"
#> [2796] "2021-10-06" "2021-10-07" "2021-10-08" "2021-10-09" "2021-10-10"
#> [2801] "2021-10-11" "2021-10-12" "2021-10-13" "2021-10-14" "2021-10-15"
#> [2806] "2021-10-16" "2021-10-17" "2021-10-18" "2021-10-19" "2021-10-20"
#> [2811] "2021-10-21" "2021-10-22" "2021-10-23" "2021-10-24" "2021-10-25"
#> [2816] "2021-10-26" "2021-10-27" "2021-10-28" "2021-10-29" "2021-10-30"
#> [2821] "2021-11-01" "2021-11-02" "2021-11-03" "2021-11-04" "2021-11-05"
#> [2826] "2021-11-06" "2021-11-07" "2021-11-08" "2021-11-09" "2021-11-10"
#> [2831] "2021-11-11" "2021-11-12" "2021-11-13" "2021-11-14" "2021-11-15"
#> [2836] "2021-11-16" "2021-11-17" "2021-11-18" "2021-11-19" "2021-11-20"
#> [2841] "2021-11-21" "2021-11-22" "2021-11-23" "2021-11-24" "2021-11-25"
#> [2846] "2021-11-26" "2021-11-27" "2021-11-28" "2021-11-29" "2021-11-30"
#> [2851] "2021-12-01" "2021-12-02" "2021-12-03" "2021-12-04" "2021-12-05"
#> [2856] "2021-12-06" "2021-12-07" "2021-12-08" "2021-12-09" "2021-12-10"
#> [2861] "2021-12-11" "2021-12-12" "2021-12-13" "2021-12-14" "2021-12-15"
#> [2866] "2021-12-16" "2021-12-17" "2021-12-18" "2021-12-19" "2021-12-20"
#> [2871] "2021-12-21" "2021-12-22" "2021-12-23" "2021-12-24" "2021-12-25"
#> [2876] "2021-12-26" "2021-12-27" "2021-12-28" "2021-12-29" "2021-12-30"
#> [2881] "2022-01-01" "2022-01-02" "2022-01-03" "2022-01-04" "2022-01-05"
#> [2886] "2022-01-06" "2022-01-07" "2022-01-08" "2022-01-09" "2022-01-10"
#> [2891] "2022-01-11" "2022-01-12" "2022-01-13" "2022-01-14" "2022-01-15"
#> [2896] "2022-01-16" "2022-01-17" "2022-01-18" "2022-01-19" "2022-01-20"
#> [2901] "2022-01-21" "2022-01-22" "2022-01-23" "2022-01-24" "2022-01-25"
#> [2906] "2022-01-26" "2022-01-27" "2022-01-28" "2022-01-29" "2022-01-30"
#> [2911] "2022-02-01" "2022-02-02" "2022-02-03" "2022-02-04" "2022-02-05"
#> [2916] "2022-02-06" "2022-02-07" "2022-02-08" "2022-02-09" "2022-02-10"
#> [2921] "2022-02-11" "2022-02-12" "2022-02-13" "2022-02-14" "2022-02-15"
#> [2926] "2022-02-16" "2022-02-17" "2022-02-18" "2022-02-19" "2022-02-20"
#> [2931] "2022-02-21" "2022-02-22" "2022-02-23" "2022-02-24" "2022-02-25"
#> [2936] "2022-02-26" "2022-02-27" "2022-02-28" "2022-02-29" "2022-02-30"
#> [2941] "2022-03-01" "2022-03-02" "2022-03-03" "2022-03-04" "2022-03-05"
#> [2946] "2022-03-06" "2022-03-07" "2022-03-08" "2022-03-09" "2022-03-10"
#> [2951] "2022-03-11" "2022-03-12" "2022-03-13" "2022-03-14" "2022-03-15"
#> [2956] "2022-03-16" "2022-03-17" "2022-03-18" "2022-03-19" "2022-03-20"
#> [2961] "2022-03-21" "2022-03-22" "2022-03-23" "2022-03-24" "2022-03-25"
#> [2966] "2022-03-26" "2022-03-27" "2022-03-28" "2022-03-29" "2022-03-30"
#> [2971] "2022-04-01" "2022-04-02" "2022-04-03" "2022-04-04" "2022-04-05"
#> [2976] "2022-04-06" "2022-04-07" "2022-04-08" "2022-04-09" "2022-04-10"
#> [2981] "2022-04-11" "2022-04-12" "2022-04-13" "2022-04-14" "2022-04-15"
#> [2986] "2022-04-16" "2022-04-17" "2022-04-18" "2022-04-19" "2022-04-20"
#> [2991] "2022-04-21" "2022-04-22" "2022-04-23" "2022-04-24" "2022-04-25"
#> [2996] "2022-04-26" "2022-04-27" "2022-04-28" "2022-04-29" "2022-04-30"
#> [3001] "2022-05-01" "2022-05-02" "2022-05-03" "2022-05-04" "2022-05-05"
#> [3006] "2022-05-06" "2022-05-07" "2022-05-08" "2022-05-09" "2022-05-10"
#> [3011] "2022-05-11" "2022-05-12" "2022-05-13" "2022-05-14" "2022-05-15"
#> [3016] "2022-05-16" "2022-05-17" "2022-05-18" "2022-05-19" "2022-05-20"
#> [3021] "2022-05-21" "2022-05-22" "2022-05-23" "2022-05-24" "2022-05-25"
#> [3026] "2022-05-26" "2022-05-27" "2022-05-28" "2022-05-29" "2022-05-30"
#> [3031] "2022-06-01" "2022-06-02" "2022-06-03" "2022-06-04" "2022-06-05"
#> [3036] "2022-06-06" "2022-06-07" "2022-06-08" "2022-06-09" "2022-06-10"
#> [3041] "2022-06-11" "2022-06-12" "2022-06-13" "2022-06-14" "2022-06-15"
#> [3046] "2022-06-16" "2022-06-17" "2022-06-18" "2022-06-19" "2022-06-20"
#> [3051] "2022-06-21" "2022-06-22" "2022-06-23" "2022-06-24" "2022-06-25"
#> [3056] "2022-06-26" "2022-06-27" "2022-06-28" "2022-06-29" "2022-06-30"
#> [3061] "2022-07-01" "2022-07-02" "2022-07-03" "2022-07-04" "2022-07-05"
#> [3066] "2022-07-06" "2022-07-07" "2022-07-08" "2022-07-09" "2022-07-10"
#> [3071] "2022-07-11" "2022-07-12" "2022-07-13" "2022-07-14" "2022-07-15"
#> [3076] "2022-07-16" "2022-07-17" "2022-07-18" "2022-07-19" "2022-07-20"
#> [3081] "2022-07-21" "2022-07-22" "2022-07-23" "2022-07-24" "2022-07-25"
#> [3086] "2022-07-26" "2022-07-27" "2022-07-28" "2022-07-29" "2022-07-30"
#> [3091] "2022-08-01" "2022-08-02" "2022-08-03" "2022-08-04" "2022-08-05"
#> [3096] "2022-08-06" "2022-08-07" "2022-08-08" "2022-08-09" "2022-08-10"
#> [3101] "2022-08-11" "2022-08-12" "2022-08-13" "2022-08-14" "2022-08-15"
#> [3106] "2022-08-16" "2022-08-17" "2022-08-18" "2022-08-19" "2022-08-20"
#> [3111] "2022-08-21" "2022-08-22" "2022-08-23" "2022-08-24" "2022-08-25"
#> [3116] "2022-08-26" "2022-08-27" "2022-08-28" "2022-08-29" "2022-08-30"
#> [3121] "2022-09-01" "2022-09-02" "2022-09-03" "2022-09-04" "2022-09-05"
#> [3126] "2022-09-06" "2022-09-07" "2022-09-08" "2022-09-09" "2022-09-10"
#> [3131] "2022-09-11" "2022-09-12" "2022-09-13" "2022-09-14" "2022-09-15"
#> [3136] "2022-09-16" "2022-09-17" "2022-09-18" "2022-09-19" "2022-09-20"
#> [3141] "2022-09-21" "2022-09-22" "2022-09-23" "2022-09-24" "2022-09-25"
#> [3146] "2022-09-26" "2022-09-27" "2022-09-28" "2022-09-29" "2022-09-30"
#> [3151] "2022-10-01" "2022-10-02" "2022-10-03" "2022-10-04" "2022-10-05"
#> [3156] "2022-10-06" "2022-10-07" "2022-10-08" "2022-10-09" "2022-10-10"
#> [3161] "2022-10-11" "2022-10-12" "2022-10-13" "2022-10-14" "2022-10-15"
#> [3166] "2022-10-16" "2022-10-17" "2022-10-18" "2022-10-19" "2022-10-20"
#> [3171] "2022-10-21" "2022-10-22" "2022-10-23" "2022-10-24" "2022-10-25"
#> [3176] "2022-10-26" "2022-10-27" "2022-10-28" "2022-10-29" "2022-10-30"
#> [3181] "2022-11-01" "2022-11-02" "2022-11-03" "2022-11-04" "2022-11-05"
#> [3186] "2022-11-06" "2022-11-07" "2022-11-08" "2022-11-09" "2022-11-10"
#> [3191] "2022-11-11" "2022-11-12" "2022-11-13" "2022-11-14" "2022-11-15"
#> [3196] "2022-11-16" "2022-11-17" "2022-11-18" "2022-11-19" "2022-11-20"
#> [3201] "2022-11-21" "2022-11-22" "2022-11-23" "2022-11-24" "2022-11-25"
#> [3206] "2022-11-26" "2022-11-27" "2022-11-28" "2022-11-29" "2022-11-30"
#> [3211] "2022-12-01" "2022-12-02" "2022-12-03" "2022-12-04" "2022-12-05"
#> [3216] "2022-12-06" "2022-12-07" "2022-12-08" "2022-12-09" "2022-12-10"
#> [3221] "2022-12-11" "2022-12-12" "2022-12-13" "2022-12-14" "2022-12-15"
#> [3226] "2022-12-16" "2022-12-17" "2022-12-18" "2022-12-19" "2022-12-20"
#> [3231] "2022-12-21" "2022-12-22" "2022-12-23" "2022-12-24" "2022-12-25"
#> [3236] "2022-12-26" "2022-12-27" "2022-12-28" "2022-12-29" "2022-12-30"