Run Model Simulations for Each Parameter Set

This function runs the epidemiological model using different parameter sets sampled from the MCMC posterior.

Usage

run_mcmc_simulations(
  model,
  param_inputs,
  param_samples,
  start_date,
  end_date,
  prewarm_years = 2,
  days_per_year = 360
)

Arguments

model

A function representing the epidemiological model.

param_samples

A matrix of parameter sets sampled from the MCMC posterior.

start_date

The simulation start date.

end_date

The simulation end date.

prewarm_years

The number of years to prewarm the model.

days_per_year

The number of days in a simulation year.

Value

A list of data frames containing simulation results for each sampled parameter set.

Examples

simulations <- run_mcmc_simulations(model, sampled_params, "2014-01-01", "2022-12-31", 2, 360)
#> Error in 1:nrow(param_samples): argument of length 0