View Default Priors

This function displays the default priors and their details, including initial values, min/max bounds, and prior distributions.

Usage

view_priors(param_inputs, proposal_matrix, params_to_estimate, priors = NULL)

Arguments

params_to_estimate

Character vector specifying which parameters' priors should be viewed.

priors

List of priors to view. If NULL, the default priors will be displayed.

Value

A data frame containing the details of each prior specified in params_to_estimate.

Examples

view_priors(param_inputs, proposal_matrix, params_to_estimate = c("a_R", "b_R", "qR", "z", "eff_SMC", "phi", "size"))
#> Error in initialize_priors(param_inputs, proposal_matrix, params_to_estimate): could not find function "initialize_priors"